% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloy.R
\docType{data}
\name{alloy}
\alias{alloy}
\title{Fatigue life for alloy T7989 specimens}
\format{
A tibble with 72 rows and 2 variables:
\describe{
  \item{cycles}{Number of cycles (in thousands).}
  \item{status}{If specimen failed before 300 thousand cycles \code{1} else \code{0}.}
}
}
\source{
Meeker, William Q; Escobar, Luis A., Statistical methods for
reliability data, New York: Wiley series in probability and statistics
(1998, p.131)
}
\usage{
alloy
}
\description{
A dataset containing the number of cycles of fatigue life for alloy T7987
specimens.
}
\keyword{datasets}
