% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{confint_fisher.parameter_estimation}
\alias{confint_fisher.parameter_estimation}
\title{Fisher Confidence Bounds for Quantiles and/or Probabilities}
\usage{
\method{confint_fisher}{parameter_estimation}(
  x,
  b_lives = c(0.01, 0.1, 0.5),
  bounds = c("two_sided", "lower", "upper"),
  conf_level = 0.95,
  direction = c("y", "x")
)
}
\arguments{
\item{x}{Object of class \code{parameter_estimation} returned from
\code{\link{ml_estimation}}.}

\item{bounds}{A character string specifying the interval(s) which has/have to
be computed. Must be one of "two_sided" (default), "lower" or "upper".}

\item{conf_level}{Confidence level of the interval. The default value is
\code{conf_level = 0.95}.}

\item{direction}{A character string specifying the direction of the computed
interval(s). Must be either "y" (failure probabilities) or "x" (quantiles).}
}
\value{
A tibble containing the lifetime characteristic, the
  probabilities, estimated standard errors by the delta method and computed
  values for the specified confidence bound(s).
}
\description{
This method computes normal-approximation confidence intervals for quantiles
and/or failure probabilities using the \code{\link{delta_method}}. The
required (log-)location-scale (and threshold) parameters and variance-covariance matrix
of these need to be estimated by Maximum Likelihood.
}
\examples{
obs   <- seq(10000, 100000, 10000)
status <- c(0, 1, 1, 0, 0, 0, 1, 0, 1, 0)
data <- reliability_data(x = obs, status = status)

tbl_john <- estimate_cdf(data, "johnson")

mle <- ml_estimation(
  data,
  distribution = "weibull",
  conf_level = 0.95
)

conf_fish <- confint_fisher(
  mle,
  distribution = "weibull",
  bounds = "two_sided",
  conf_level = 0.95,
  direction = "y"
)
}
