% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_distributions.R
\name{dist_delay_report}
\alias{dist_delay_report}
\title{Parameter estimation of the delay in report distribution}
\usage{
dist_delay_report(date_repair, date_report, distribution = "lognormal")
}
\arguments{
\item{date_repair}{a vector of class \code{"character"} or \code{"Date"}
indicating the date of repair of a failed unit.}

\item{date_report}{a vector of class \code{"character"} or \code{"Date"}
indicating the date of the report of a failed unit.}

\item{distribution}{supposed distribution of the random variable. The default
value is \code{"lognormal"}.}
}
\value{
A named vector of estimated parameters for the specified
  distribution.
}
\description{
This function generates a delay random variable by calculating the time
difference between the report and repair date for the sample units
and afterwards estimates the parameter(s) of a supposed distribution,
using MLE.
}
\examples{
date_of_repair <- c("2014-10-21", "2014-09-15", "2015-07-04", "2015-04-10",
                    "2015-02-15", "2015-04-14", "2015-04-24", "2015-02-27",
                    "2015-04-25", "2015-04-24", "2015-06-12", "2015-08-26",
                    "2015-05-04", "2015-04-04", "2015-09-06", "2015-05-22",
                    "2015-08-21", "2015-09-17", "2015-09-15", "2015-08-15",
                    "2015-11-26", "2015-08-22", "2015-10-05")
date_of_report <- c("2014-11-14", "2014-10-09", "2015-08-28", "2015-04-15",
                    "2015-02-28", "2015-07-03", "2015-05-16", "2015-03-08",
                    "2015-05-28", "2015-05-15", "2015-07-11", "2015-09-12",
                    "2015-08-14", "2015-05-06", "2015-09-10", "2015-06-05",
                    "2015-09-03", "2015-10-17", "2015-10-01", "2015-08-21",
                    "2015-12-02", "2015-08-28", "2015-11-01")

params_delay_report  <- dist_delay_report(date_repair = date_of_repair,
                                            date_report = date_of_report,
                                            distribution = "lognormal")
}
