% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{estimate_cdf}
\alias{estimate_cdf}
\title{Estimation of Failure Probabilities}
\usage{
estimate_cdf(data, ...)
}
\arguments{
\item{data}{A standardised data frame returned by \link{reliability_data}.}

\item{method}{Method used for estimating the failure probablities. See 'Details'.}

\item{options}{A list of named options passed to \code{<method>_method}. For
now there is just the option "method" with \code{method = "mr"}. See
\code{\link{mr_method}}.}
}
\description{
Estimate failure probabilities for uncensored or (multiple) right-censored
data.
}
\details{
Depending on the \code{method} argument one of the following functions is
called:
\itemize{
  \item \code{"mr"}: \code{\link{mr_method}}. Works only for uncensored data.
  \item \code{"johnson"}: \code{\link{johnson_method}}.
  \item \code{"kaplan"}: \code{\link{kaplan_method}}.
  \item \code{"nelson"}: \code{\link{nelson_method}}.
}
}
