% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/failure_data.R
\name{failure_data}
\alias{failure_data}
\title{Failure data}
\usage{
failure_data(data = NULL, x, status, id)
}
\arguments{
\item{data}{Either NULL or a data frame/tibble. If data is NULL, x, status and
id must be vectors containing the data. Otherwise x, status and id can be either
column names or column positions.}

\item{x}{Lifetime data, that means any characteristic influencing the reliability
of a product, e.g. operating time (days/months in service), mileage (km, miles),
load cycles.}

\item{status}{Binary data (0 or 1) indicating whether a unit is a right
censored observation (= 0) or a failure (= 1).}

\item{id}{Identification for every unit.}
}
\value{
A tibble with class \code{"failure_data"} containing the following
columns:
\itemize{
  \item \code{x} Lifetime characteristic.
  \item \code{status} Binary data (0 or 1) indicating whether a unit is a right
    censored observation (= 0) or a failure (= 1).
  \item \code{id} Identification for every unit.
}
}
\description{
Create consistent failure data based on an existing data.frame/tibble
(preferred) or on multiple equal length vectors.
}
\examples{
# Alloy T7987 dataset taken from Meeker and Escobar (1998, p. 131)
cycles   <- c(300, 300, 300, 300, 300, 291, 274, 271, 269, 257, 256, 227, 226,
              224, 213, 211, 205, 203, 197, 196, 190, 189, 188, 187, 184, 180,
              180, 177, 176, 173, 172, 171, 170, 170, 169, 168, 168, 162, 159,
              159, 159, 159, 152, 152, 149, 149, 144, 143, 141, 141, 140, 139,
              139, 136, 135, 133, 131, 129, 123, 121, 121, 118, 117, 117, 114,
              112, 108, 104, 99, 99, 96, 94)
state <- c(rep(0, 5), rep(1, 67))
id <- "XXXXXX"

Example 1: Based on existing data.frame/tibble
tbl <- tibble::tibble(x = cycles, status = state, id = id)
failure_tbl <- failure_data(tbl, x = x, status = status, id = id)

Example 2: Based on vectors
failure_tbl_2 <- failure_data(x = cycles, status = state, id = id)

}
