% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\encoding{UTF-8}
\name{loglik_function.default}
\alias{loglik_function.default}
\title{Log-Likelihood Function for Parametric Lifetime Distributions}
\usage{
\method{loglik_function}{default}(
  x,
  status,
  wts = rep(1, length(x)),
  dist_params,
  distribution = c("weibull", "lognormal", "loglogistic", "sev", "normal", "logistic",
    "weibull3", "lognormal3", "loglogistic3"),
  ...
)
}
\arguments{
\item{x}{A numeric vector which consists of lifetime data. Lifetime data
could be every characteristic influencing the reliability of a product,
e.g. operating time (days/months in service), mileage (km, miles), load
cycles.}

\item{status}{A vector of binary data (0 or 1) indicating whether a unit is
a right censored observation (= 0) or a failure (= 1).}

\item{wts}{Optional vector of case weights. The length of \code{wts} must be equal
to the number of observations in \code{x}.}

\item{dist_params}{A vector of parameters. An overview of the
distribution-specific parameters can be found in section 'Distributions'.}

\item{distribution}{Supposed distribution of the random variable.}

\item{...}{Further arguments passed to or from other methods. Currently not used.}
}
\value{
Returns the log-likelihood value for the parameters in \code{dist_params} given
the data.
}
\description{
This function computes the log-likelihood value with respect to a given set
of parameters. For two-parametric models the location and scale parameters
are required. If a three-parametric lifetime distribution is needed an
additional threshold parameter has to be provided. In terms of
\emph{Maximum Likelihood Estimation} this function can be optimized
(\link[stats:optim]{optim}) to estimate the parameters and variance-covariance
matrix of the parameters.
}
\section{Distributions}{

The following table summarizes the available distributions and their parameters
\itemize{
\item \emph{location parameter} \eqn{\mu},
\item \emph{scale parameter} \eqn{\sigma} or \eqn{\theta} and
\item \emph{threshold parameter} \eqn{\gamma}.
}

The order within \code{dist_params} is given in the table header.\tabular{lccc}{
   \strong{\code{distribution}} \tab \strong{\code{dist_params[1]}} \tab \strong{\code{dist_params[2]}} \tab \strong{\code{dist_params[3]}} \cr
   \code{"sev"} \tab \eqn{\mu} \tab \eqn{\sigma} \tab - \cr
   \code{"weibull"} \tab \eqn{\mu} \tab \eqn{\sigma} \tab - \cr
   \code{"weibull3"} \tab \eqn{\mu} \tab \eqn{\sigma} \tab \eqn{\gamma} \cr
   \code{"normal"} \tab \eqn{\mu} \tab \eqn{\sigma} \tab - \cr
   \code{"lognormal"} \tab \eqn{\mu} \tab \eqn{\sigma} \tab - \cr
   \code{"lognormal3"} \tab \eqn{\mu} \tab \eqn{\sigma} \tab \eqn{\gamma} \cr
   \code{"logistic"} \tab \eqn{\mu} \tab \eqn{\sigma} \tab - \cr
   \code{"loglogistic"} \tab \eqn{\mu} \tab \eqn{\sigma} \tab - \cr
   \code{"loglogistic3"} \tab \eqn{\mu} \tab \eqn{\sigma} \tab \eqn{\gamma} \cr
   \code{"exponential"} \tab \eqn{\theta} \tab - \tab - \cr
   \code{"exponential2"} \tab \eqn{\theta} \tab \eqn{\gamma} \tab - \cr
}
}

\examples{
# Vectors:
cycles <- alloy$cycles
status <- alloy$status

# Example 1 - Evaluating Log-Likelihood function of two-parametric weibull:
loglik_weib <- loglik_function(
  x = cycles,
  status = status,
  dist_params = c(5.29, 0.33),
  distribution = "weibull"
)

# Example 2 - Evaluating Log-Likelihood function of three-parametric weibull:
loglik_weib3 <- loglik_function(
  x = cycles,
  status = status,
  dist_params = c(4.54, 0.76, 92.99),
  distribution = "weibull3"
)

}
\references{
Meeker, William Q; Escobar, Luis A., Statistical methods for
reliability data, New York: Wiley series in probability and statistics, 1998
}
\seealso{
\link{loglik_function}
}
