% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_distributions.R
\name{mcs_delay}
\alias{mcs_delay}
\title{Correction of Operating Times by Delays using a Monte Carlo Approach}
\usage{
mcs_delay(
  date_1,
  date_2,
  x,
  status,
  distribution = c("lognormal", "exponential"),
  seed = NULL
)
}
\arguments{
\item{date_1}{A vector of class \code{"character"} or \code{"Date"}, in the
format "yyyy-mm-dd", indicating the earlier of the two dates. If no date is
available use \code{NA}.}

\item{date_2}{A vector of class \code{"character"} or \code{"Date"}, in the
format "yyyy-mm-dd", indicating the later of the two dates. If no date is
available use \code{NA}.}

\item{x}{A numeric vector of operating times.}

\item{status}{A vector of binary data (0 or 1) indicating whether unit \emph{i}
is a right censored observation (= 0) or a failure (= 1).}

\item{distribution}{Supposed distribution of the random variable. The default
value is \code{"lognormal"}.}

\item{seed}{If \code{seed = NULL} a random seed is used. Otherwise the user
can specify an integer for the seed.}
}
\value{
A numeric vector of corrected operating times for the censored units
  and the input operating times for the failed units if
  \code{details = FALSE}. If \code{details = TRUE} the output is a list which
  consists of the following elements:
  \itemize{
  \item \code{time} : Numeric vector of corrected operating times for the
    censored observations and input operating times for failed units.
  \item \code{x_sim} : Simulated random numbers of specified distribution with
    estimated parameters. The length of \code{x_sim} is equal to the number of
    censored observations.
  \item \code{coefficients} : Estimated coefficients of supposed
    distribution.
  \item \code{int_seed} : Integer seed number for reproducibility.}
}
\description{
In general, the amount of available data about units in the field is very
different. While there are many information on failures during the warranty
period, little is known about the units without a failure (\emph{censored data}).
As a result, the operating times of these units are often inaccurate and must
be adjusted by delays. But usually delays are only known for defective units.
See 'Details' for more information.

This function reduces the operating times of (multiple) right censored observations
by simulated delays which were drawn from the distribution determined by
complete data (described in 'Details' of \code{\link{dist_delay}}).
}
\details{
In field data analysis time-dependent characteristics (e.g. \emph{time in service})
are often imprecisely recorded. These inaccuracies are caused by unconsidered delays.

For a better understanding of the MCS application in the context of field data,
two cases are described below.
\itemize{
  \item \strong{Delay in registration}: It is common that a supplier, who provides
    parts to the manufacturing industry does not know when the unit, in which
    its parts are installed, were put in service (due to unknown registration
    or sales date). Without taking the described delay into account, the time
    in service of intact units would be the difference between the present
    date and the production date. But the actual operating time is (much) shorter,
    since the stress on the component will not start until the complete system
    is put in service. Therefore the intact units must be reduced by delays.
  \item \strong{Delay in report}: Authorized repairers often do not immediately
    notify the manufacturer or OEM of repairs that occur during the warranty
    period, but instead pass the information about these repairs in form of
    collected applications e.g. monthly or quarterly. The resulting time
    difference between the reporting of the repair in the guarantee database
    and the actual repair date is called the reporting delay. For a given date
    where the analysis is made there could be units which had a failure but are
    not registered therefore treated as censored units. In order to take this
    case into account and according to the principle of equal
    opportunities, the lifetime of intact units is reduced by simulated
    reporting delays, so that the case described before is taken into account
    due to lifetime reduction.
}
}
\examples{
date_of_production   <- c("2014-07-28", "2014-02-17", "2014-07-14",
                          "2014-06-26", "2014-03-10", "2014-05-14",
                          "2014-05-06", "2014-03-07", "2014-03-09",
                          "2014-04-13", "2014-05-20", "2014-07-07",
                          "2014-01-27", "2014-01-30", "2014-03-17",
                          "2014-02-09", "2014-04-14", "2014-04-20",
                          "2014-03-13", "2014-02-23", "2014-04-03",
                          "2014-01-08", "2014-01-08")
date_of_registration <- c(NA, "2014-03-29", "2014-12-06", "2014-09-09",
                          NA, NA, "2014-06-16", NA, "2014-05-23",
                          "2014-05-09", "2014-05-31", NA, "2014-04-13",
                          NA, NA, "2014-03-12", NA, "2014-06-02",
                          NA, "2014-03-21", "2014-06-19", NA, NA)

op_time <- rep(1000, length(date_of_production))
state <- c(0, 1, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 0)

# Example 1 - MCS for delay in registration:
list_detail <- mcs_delay(
  date_1 = date_of_production,
  date_2 = date_of_registration,
  x = op_time,
  status = state,
  distribution = "lognormal",
  seed = NULL
)
}
\references{
Verband der Automobilindustrie e.V. (VDA); Qualitätsmanagement in
  der Automobilindustrie. Zuverlässigkeitssicherung bei Automobilherstellern
  und Liferanten. Zuverlässigkeits-Methoden und -Hilfsmittel.; 4th Edition, 2016,
  <ISSN:0943-9412>
}
