% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_identification.R
\encoding{UTF-8}
\name{mixmod_em}
\alias{mixmod_em}
\title{Mixture Model Estimation using EM-Algorithm}
\usage{
mixmod_em(x, ...)
}
\value{
Returns a list where the length of the list depends on the number of
   k subgroups. The first \code{k} lists have the same information as provided by
   \link{ml_estimation}, but the values \code{logL}, \code{aic} and \code{bic} are
   the results of a log-likelihood function, which is weighted by a-posteriori
   probabilities. The last list summarizes further results of the EM-Algorithm and
   is therefore called \code{em_results}. It contains the following elements:
  \itemize{
  \item \code{a_priori} : A vector with estimated a-priori probabilities.
  \item \code{a_posteriori} : A matrix with estimated a-posteriori probabilities.
  \item \code{groups} : Numeric vector specifying the group membership of every
    observation.
  \item \code{logL} : The value of the complete log-likelihood.
  \item \code{aic} : Akaike Information Criterion.
  \item \code{bic} : Bayesian Information Criterion.}
}
\description{
This method uses the EM-Algorithm to estimate the parameters of a univariate
mixture model. Until now, the mixture model can consist of k two-parametric
Weibull distributions. If no mixture of k components can be estimated,
the function is forced to stop and a message with instructions is given.
}
\details{
In \code{mixmod_em} the function \code{\link{mixture_em_cpp}} is called. The
computed posterior probabilities are then used as weights inside function
\code{\link{ml_estimation}} to model a weighted log-likelihood. This strategy
enables the computation of confidence intervals for the parameters of the
separated sub-distributions, since \code{ml_estimation} provides a variance-covariance
matrix. Using this strategy, a potential problem that can occur is,
that the value of the complete log-likelihood, computed by \code{mixture_em_cpp},
differs considerably from the complete log-likelihood after re-estimating
parameters with \code{ml_estimation}. If so, the estimated quantities like
prior and posterior probabilities, as well as the model parameters are not
reliable anymore and the function is forced to stop with the message:
"Parameter estimation was not successful!"
But if the log-likelihood values are close to each other, the presence of the
mixture is strengthened and a reasonable fit is provided.
Thus, a check of the absolute differences in the log-likelihood values is made
and the critical difference has to be specified in argument \code{diff_loglik}.
}
\section{Methods (by class)}{

\describe{
  \item{\code{\link[=mixmod_em.reliability_data]{reliability_data}}}{
    Preferred. Provide the output of \code{\link{reliability_data}} directly.
  }
  \item{\code{\link[=mixmod_em.default]{default}}}{
    Provide \code{x} and \code{status} manually.
  }
}
}

\references{
\itemize{
    \item Doganaksoy, N.; Hahn, G.; Meeker, W. Q., Reliability Analysis by
      Failure Mode, Quality Progress, 35(6), 47-52, 2002
    \item Blog posts by Stefan Gelissen: \url{http://blogs2.datall-analyse.nl/2016/02/18/rcode_mixture_distribution_censored};
      last access on 19th January 2019}
}
