% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_identification.R
\name{mixmod_regression.cdf_estimation}
\alias{mixmod_regression.cdf_estimation}
\title{Mixture Model Identification using Segmented Regression}
\usage{
\method{mixmod_regression}{cdf_estimation}(
  x,
  distribution = c("weibull", "lognormal", "loglogistic"),
  conf_level = 0.95,
  control = segmented::seg.control(),
  ...
)
}
\arguments{
\item{x}{An object of class \code{cdf_estimation} returned from
\code{\link{estimate_cdf}}.}

\item{distribution}{Supposed distribution of the random variable.}

\item{conf_level}{Confidence level of the interval. If \code{distribution} is
\code{"weibull"} this must be one of \code{0.9}, \code{0.95} or \code{0.99}.}

\item{control}{Output of call to \code{\link[segmented]{seg.control}}, which
is passed to \code{\link[segmented:segmented.lm]{segmented.lm}}.}
}
\value{
Returns a list where the length of the list depends on
  the number of identified subgroups. Each list has the same
  information as provided by \link{rank_regression}. Additionally each list
  has an element that specifies the range regarding the lifetime data for
  every subgroup.
}
\description{
This method uses piecewise linear regression to separate the data in
subgroups, if appropriate. Since this happens in an automated fashion
the function tends to overestimate the number of breakpoints and
therefore returns too many subgroups. This problem is already stated in
the documentation of the function \link{segmented.lm}, which is part of
the \emph{segmented} package. A maximum of three subgroups can be obtained.
}
\examples{
# Data is taken from given reference:
hours <- c(2, 28, 67, 119, 179, 236, 282, 317, 348, 387, 3, 31, 69, 135,
          191, 241, 284, 318, 348, 392, 5, 31, 76, 144, 203, 257, 286,
          320, 350, 412, 8, 52, 78, 157, 211, 261, 298, 327, 360, 446,
          13, 53, 104, 160, 221, 264, 303, 328, 369, 21, 64, 113, 168,
          226, 278, 314, 328, 377)

status <- c(1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 0, 1,
          1, 0, 1, 1, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 0,
          1, 1, 1, 1, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1,
          0, 1, 1, 1, 1, 1, 1)

data <- reliability_data(x = hours, status = status)

tbl_john <- estimate_cdf(data, methods = "johnson")

mix_mod <- mixmod_regression(tbl_john, distribution = "weibull")

}
\references{
Doganaksoy, N.; Hahn, G.; Meeker, W. Q., Reliability Analysis by
  Failure Mode, Quality Progress, 35(6), 47-52, 2002
}
