% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_identification.R
\encoding{UTF-8}
\name{mixmod_regression}
\alias{mixmod_regression}
\title{Mixture Model Identification using Segmented Regression}
\usage{
mixmod_regression(x, ...)
}
\value{
Returns a list where the length of the list depends on
  the number of identified subgroups. Each list has the same
  information as provided by \link{rank_regression}. Additionally each list
  has an element that specifies the range regarding the lifetime data for
  every subgroup.
}
\description{
This method uses piecewise linear regression to separate the data in
subgroups, if appropriate. Since this happens in an automated fashion
the function tends to overestimate the number of breakpoints and
therefore returns too many subgroups. This problem is already stated in
the documentation of the function \link{segmented.lm}, which is part of
the \emph{segmented} package. A maximum of three subgroups can be obtained.
}
\section{Methods (by class)}{

\describe{
  \item{\code{\link[=mixmod_regression.cdf_estimation]{cdf_estimation}}}{
    Preferred. Provide the output of \code{\link{estimate_cdf}} directly.
  }
  \item{\code{\link[=mixmod_regression.default]{default}}}{
    Provide \code{x}, \code{y} and \code{event} manually.
  }
}
}

\references{
Doganaksoy, N.; Hahn, G.; Meeker, W. Q., Reliability Analysis by
  Failure Mode, Quality Progress, 35(6), 47-52, 2002
}
