% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_estimation.R
\name{ml_estimation.default}
\alias{ml_estimation.default}
\title{ML Estimation for Parametric Lifetime Distributions}
\usage{
\method{ml_estimation}{default}(
  x,
  status,
  distribution = c("weibull", "lognormal", "loglogistic", "normal", "logistic", "sev",
    "weibull3", "lognormal3", "loglogistic3"),
  wts = rep(1, length(x)),
  conf_level = 0.95,
  ...
)
}
\arguments{
\item{x}{A numeric vector which consists of lifetime data. Lifetime
data could be every characteristic influencing the reliability of a product,
e.g. operating time (days/months in service), mileage (km, miles), load
cycles.}

\item{status}{A vector of binary data (0 or 1) indicating whether unit \emph{i}
is a right censored observation (= 0) or a failure (= 1).}

\item{distribution}{Supposed distribution of the random variable.}

\item{wts}{Optional vector of case weights. The length of \code{wts} must be the
same as the number of observations in \code{x}. Default is that \code{wts} is a
vector with all components being 1 (same weights).}

\item{conf_level}{Confidence level of the interval.}
}
\value{
Returns a list with the following elements:
  \itemize{
    \item \code{coefficients} : If \code{distribution} is \code{"weibull"}, the
      estimated scale (\eqn{\eta}) and shape (\eqn{\beta}) parameters are provided
      (and additionally the threshold parameter \eqn{\gamma} if \code{distribution}
      is \code{"weibull3"}). For any other distribution, \code{coefficients} is
      identical to \code{loc_sc_params}.
    \item \code{confint} : Only included if \code{distribution} is \code{"weibull"}
      or \code{"weibull3"}. Confidence intervals for \eqn{\eta} and \eqn{\beta}
      (and \eqn{\gamma} if \code{distribution} is \code{"weibull3"}).
    \item \code{loc_sc_params} : Estimated (log-)location-scale parameters.
      Threshold parameter is included if \code{distribution} is \code{"weibull3"},
      \code{"lognormal3"} or \code{"loglogistic3"}.
    \item \code{loc_sc_confint} : Confidence intervals the (log-)location-scale
      parameters.
    \item \code{loc_sc_varcov} : Estimated variance-covariance matrix for the
      (log-)location-scale parameters.
    \item \code{logL} : The log-likelihood value.
    \item \code{aic} : Akaike Information Criterion.
    \item \code{bic} : Bayesian Information Criterion.
    \item \code{data} : A tibble with columns \code{x} and \code{status}.
    \item \code{distribution} : Specified distribution.
  }
}
\description{
This method estimates the parameters of a two- or three-parametric lifetime
distribution for complete and (multiple) right censored data. The parameters
are determined in the frequently used (log-) location-scale
parameterization.

For the Weibull, estimates are transformed such that
they are in line with the parameterization provided by the \emph{stats} package
(see \link[stats]{Weibull}).
}
\details{
\code{ml_estimation} calls \code{\link[SPREDA:lifedata.MLE]{Lifedata.MLE}},
which is implemented in \emph{SPREDA}, to obtain the estimates. Normal
approximation confidence intervals for the parameters are computed as well.
}
\examples{
# Vectors:
obs <- seq(10000, 100000, 10000)
status_1 <- c(0, 1, 1, 0, 0, 0, 1, 0, 1, 0)

cycles <- alloy$cycles
status_2 <- alloy$status

# Example 1 - Fitting a two-parametric weibull distribution:
ml <- ml_estimation(
  x = obs,
  status = status_1,
  distribution = "weibull",
  conf_level = .90
)

# Example 2 - Fitting a three-parametric lognormal distribution:
ml_2 <- ml_estimation(
  x = cycles,
  status = status_2,
  distribution = "lognormal"
)

}
\references{
Meeker, William Q; Escobar, Luis A., Statistical methods for
  reliability data, New York: Wiley series in probability and statistics, 1998
}
\seealso{
\code{\link{ml_estimation}}
}
