% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_estimation.R
\name{ml_estimation.reliability_data}
\alias{ml_estimation.reliability_data}
\title{ML Estimation for Parametric Lifetime Distributions}
\usage{
\method{ml_estimation}{reliability_data}(
  x,
  distribution = c("weibull", "lognormal", "loglogistic", "normal", "logistic", "sev",
    "weibull3", "lognormal3", "loglogistic3"),
  wts = rep(1, nrow(x)),
  conf_level = 0.95,
  ...
)
}
\arguments{
\item{x}{An object of class \code{reliability_data} returned by
\code{\link{reliability_data}}.}

\item{distribution}{Supposed distribution of the random variable.}

\item{wts}{Optional vector of case weights. The length of \code{wts} must be the
same as the number of observations \code{x}. Default is that \code{wts} is a
vector with all components being 1 (same weights).}

\item{conf_level}{Confidence level of the interval. The default value is
\code{conf_level = 0.95}.}
}
\value{
Returns a list with the following components:
  \itemize{
  \item \code{coefficients} : Provided, if \code{distribution} is \code{"weibull"}.
    \eqn{\eta} is the estimated scale and \eqn{\beta} the estimated shape parameter.
  \item \code{confint} : Provided, if \code{distribution} is \code{"weibull"}.
    Confidence interval for \eqn{\eta} and \eqn{\beta}.
  \item \code{loc_sc_params} : Estimated location-scale parameters.
  \item \code{loc_sc_confint} : Confidence interval for location-scale parameters.
  \item \code{loc_sc_varcov} : Estimated Variance-Covariance matrix of the used
    location-scale distribution.
  \item \code{logL} : The log-likelihood value.
  \item \code{aic} : Akaike Information Criterion.
  \item \code{bic} : Bayesian Information Criterion.}
}
\description{
This method estimates the parameters and calculates normal approximation confidence
intervals for a two- or three-parametric lifetime distribution in the frequently used
(log-) location-scale parameterization. \code{ml_estimation} uses the
\code{\link{Lifedata.MLE}} function which is defined in the
\emph{SPREDA} package.
For the Weibull the estimates are transformed such that they are in line with
the parameterization provided by the \emph{stats} package like
\code{\link{pweibull}}. The method is applicable for complete and (multiple)
right censored data.
}
\examples{
# Example 1: Fitting a two-parameter Weibull:
obs   <- seq(10000, 100000, 10000)
status <- c(0, 1, 1, 0, 0, 0, 1, 0, 1, 0)
data <- reliability_data(x = obs, status = status)

mle <- ml_estimation(
  data,
  distribution = "weibull",
  conf_level = 0.90
)

# Example 2: Fitting a three-parameter Weibull:
# Alloy T7987 dataset taken from Meeker and Escobar(1998, p. 131)
cycles   <- c(300, 300, 300, 300, 300, 291, 274, 271, 269, 257, 256, 227, 226,
              224, 213, 211, 205, 203, 197, 196, 190, 189, 188, 187, 184, 180,
              180, 177, 176, 173, 172, 171, 170, 170, 169, 168, 168, 162, 159,
              159, 159, 159, 152, 152, 149, 149, 144, 143, 141, 141, 140, 139,
              139, 136, 135, 133, 131, 129, 123, 121, 121, 118, 117, 117, 114,
              112, 108, 104, 99, 99, 96, 94)
status <- c(rep(0, 5), rep(1, 67))
data_2 <- reliability_data(x = cycles, status = status)

mle_weib3 <- ml_estimation(
  data_2,
  distribution = "weibull3",
  conf_level = 0.95
)

}
\references{
Meeker, William Q; Escobar, Luis A., Statistical methods for
  reliability data, New York: Wiley series in probability and statistics, 1998
}
