% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\encoding{UTF-8}
\name{plot_conf}
\alias{plot_conf}
\alias{plot_conf.confint}
\title{Add Confidence Region(s) for Quantiles or Probabilities}
\usage{
plot_conf(p_obj, x, ...)

\method{plot_conf}{confint}(
  p_obj,
  x,
  title_trace_mod = "Fit",
  title_trace_conf = "Confidence Limit",
  ...
)
}
\arguments{
\item{p_obj}{A plot object returned from \code{\link{plot_prob}}.}

\item{x}{Confindence interval as returned by \code{\link{confint_betabinom}}
or \code{\link{confint_fisher}}.}

\item{...}{Further arguments passed to or from other methods. Currently not used.}

\item{mod}{A model object. See 'S3 methods' for valid classes.}
}
\value{
Returns a plot object containing the probability plot with
  plotting positions, the estimated regression line and the estimated
  confidence region(s).
}
\description{
This function is used to add estimated confidence region(s) to an existing
probability plot which also includes the estimated regression line.
}
\section{S3 methods}{
\tabular{lll}{
   \strong{Class} \tab \strong{Returned from} \tab \strong{Condition} \cr
   \code{model_estimation} \tab \code{\link{rank_regression}} \tab One method in \code{\link{estimate_cdf}} \cr
   \code{model_estimation} \tab \code{\link{mixmod_regression}} \tab One method in \code{\link{estimate_cdf}}. No subgroups determined. \cr
   \code{model_estimation} \tab \code{\link{ml_estimation}} \tab - \cr
   \code{model_estimation_list} \tab \code{\link{rank_regression}} \tab Multiple methods in \code{\link{estimate_cdf}} \cr
   \code{mixmod_regression} \tab \code{\link{mixmod_regression}} \tab One method in \code{\link{estimate_cdf}} \cr
   \code{mixmod_regression_list} \tab \code{\link{mixmod_regression}} \tab Multiple methods in \code{\link{estimate_cdf}} \cr
   \code{mixmod_em} \tab \code{\link{mixmod_em}} \tab - \cr
}
}

\examples{
data <- reliability_data(data = alloy, x = cycles, status = status)

prob_tbl <- estimate_cdf(data, methods = "johnson")

# Example 1 - Probability Plot, Regression Line and Confidence Bounds for Three-Parameter-Weibull:
rr <- rank_regression(
  x = prob_tbl,
  distribution = "weibull3",
  conf_level = 0.9
)

conf_betabin <- confint_betabinom(
  x = rr,
  bounds = "two_sided",
  conf_level = 0.95,
  direction = "y"
)

plot_weibull <- plot_prob(
  x = prob_tbl,
  distribution = "weibull",
  title_main = "Three-Parametric Weibull",
  title_x = "Cycles",
  title_y = "Probability of Failure in \%",
  title_trace = "Failed Items"
)

plot_conf_beta <- plot_conf(
  p_obj = plot_weibull,
  x = conf_betabin,
  title_trace_mod = "Estimated Weibull CDF",
  title_trace_conf = "Confidence Region"
)

# Example 2 - Probability Plot, Regression Line and Confidence Bounds for Three-Parameter-Lognormal:
rr_ln <- rank_regression(
  x = prob_tbl,
  distribution = "lognormal3",
  conf_level = 0.9
)

conf_betabin_ln <- confint_betabinom(
  x = rr_ln,
  bounds = "two_sided",
  conf_level = 0.9,
  direction = "y"
)

plot_lognormal <- plot_prob(
  x = prob_tbl,
  distribution = "lognormal",
  title_main = "Three-Parametric Lognormal",
  title_x = "Cycles",
  title_y = "Probability of Failure in \%",
  title_trace = "Failed Items"
)

plot_conf_beta_ln <- plot_conf(
  p_obj = plot_lognormal,
  x = conf_betabin_ln,
  title_trace_fit = "Estimated Lognormal CDF",
  title_trace_conf = "Confidence Region"
)

}
\references{
Meeker, William Q; Escobar, Luis A., Statistical methods for
  reliability data, New York: Wiley series in probability and statistics, 1998
}
