% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_mod.model_estimation}
\alias{plot_mod.model_estimation}
\alias{plot_mod.model_estimation_list}
\title{Add Estimated Population Line(s) to a Probability Plot}
\usage{
\method{plot_mod}{model_estimation}(p_obj, x, title_trace = "Fit", ...)

\method{plot_mod}{model_estimation_list}(p_obj, x, title_trace = "Fit", ...)
}
\arguments{
\item{p_obj}{A plot object returned from \code{\link{plot_prob}}.}

\item{x}{An object of class \code{model_estimation} or
\code{model_estimation_list} returned by either
\code{\link{ml_estimation}} or \code{\link{rank_regression}}.}

\item{title_trace}{A character string whis is assigned to the trace shown in
the legend.}
}
\value{
Returns a plotly object containing the probability plot with
  plotting positions and the estimated regression line.
}
\description{
This function adds regression lines to an existing probability plot using a
model estimated by \code{\link{rank_regression}} or
\code{\link{ml_estimation}}.
}
\examples{
# Alloy T7987 dataset taken from Meeker and Escobar(1998, p. 131)
cycles   <- c(300, 300, 300, 300, 300, 291, 274, 271, 269, 257, 256, 227, 226,
              224, 213, 211, 205, 203, 197, 196, 190, 189, 188, 187, 184, 180,
              180, 177, 176, 173, 172, 171, 170, 170, 169, 168, 168, 162, 159,
              159, 159, 159, 152, 152, 149, 149, 144, 143, 141, 141, 140, 139,
              139, 136, 135, 133, 131, 129, 123, 121, 121, 118, 117, 117, 114,
              112, 108, 104, 99, 99, 96, 94)
status <- c(rep(0, 5), rep(1, 67))

data <- reliability_data(x = cycles, status = status)

tbl_john <- estimate_cdf(data, methods = c("johnson", "nelson"))

# Example 1: Probability Plot and Regression Line Three-Parameter-Weibull:
plot_weibull <- plot_prob(
  tbl_john,
  distribution = "weibull",
  title_main = "Three-Parametric Weibull",
  title_x = "Cycles",
  title_y = "Probability of Failure in \%",
  title_trace = "Failed Items"
)

mrr <- rank_regression(
  tbl_john,
  distribution = "weibull3",
  conf_level = .90
)

plot_reg_weibull <- plot_mod(
  p_obj = plot_weibull,
  x = mrr,
  title_trace = "Estimated Weibull CDF"
)



# Example 2: Probability Plot and Regression Line Three-Parameter-Lognormal:
plot_lognormal <- plot_prob(
  tbl_john,
  distribution = "lognormal",
  title_main = "Three-Parametric Lognormal",
  title_x = "Cycles",
  title_y = "Probability of Failure in \%",
  title_trace = "Failed Items"
)

mrr_ln <- rank_regression(
  tbl_john,
  distribution = "lognormal3",
  conf_level = .90
)

plot_reg_lognormal <- plot_mod(
  p_obj = plot_lognormal,
  x = mrr_ln,
  title_trace = "Estimated Lognormal CDF"
)

}
\references{
Meeker, William Q; Escobar, Luis A., Statistical methods for
  reliability data, New York: Wiley series in probability and statistics, 1998
}
