% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_mod_mix.model_estimation}
\alias{plot_mod_mix.model_estimation}
\alias{plot_mod_mix.mixmod_regression_list}
\alias{plot_mod_mix.mixmod_em}
\title{Adding Estimated Population Lines of a Separated Mixture Model to a
Probability Plot}
\usage{
\method{plot_mod_mix}{model_estimation}(p_obj, x, title_trace = "Fit", ...)

\method{plot_mod_mix}{mixmod_regression_list}(p_obj, x, title_trace = "Fit", ...)

\method{plot_mod_mix}{mixmod_em}(p_obj, x, title_trace = "Fit", ...)
}
\description{
Adding Estimated Population Lines of a Separated Mixture Model to a
Probability Plot
}
