% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_prob.cdf_estimation}
\alias{plot_prob.cdf_estimation}
\title{Probability Plotting Method for Univariate Lifetime Distributions}
\usage{
\method{plot_prob}{cdf_estimation}(
  x,
  distribution = c("weibull", "lognormal", "loglogistic", "normal", "logistic", "sev"),
  title_main = "Probability Plot",
  title_x = "Characteristic",
  title_y = "Unreliability",
  title_trace = "Sample",
  plot_method = c("plotly", "ggplot2")
)
}
\arguments{
\item{x}{CDF estimation returned from \code{\link{estimate_cdf}}.}

\item{distribution}{Supposed distribution of the random variable.}

\item{title_main}{A character string which is assigned to the main title
of the plot.}

\item{title_x}{A character string which is assigned to the title of the
x axis.}

\item{title_y}{A character string which is assigned to the title of the
y axis.}

\item{title_trace}{A character string whis is assigned to the trace shown in
the legend.}

\item{plot_method}{Package, which is used for generating the plot output.}
}
\value{
Returns a plotly object containing the layout of the probability plot
  provided by \code{\link{plot_layout}} and the plotting positions.
}
\description{
This function is used to apply the graphical technique of probability
plotting.
}
\details{
The marker label for x and y are determined by the first word provided in the
argument \code{title_x} respective \code{title_y}, i.e. if
\code{title_x = "Mileage in km"} the x label of the marker is "Mileage".
}
\examples{
# Alloy T7987 dataset taken from Meeker and Escobar(1998, p. 131)
cycles   <- c(300, 300, 300, 300, 300, 291, 274, 271, 269, 257, 256, 227, 226,
              224, 213, 211, 205, 203, 197, 196, 190, 189, 188, 187, 184, 180,
              180, 177, 176, 173, 172, 171, 170, 170, 169, 168, 168, 162, 159,
              159, 159, 159, 152, 152, 149, 149, 144, 143, 141, 141, 140, 139,
              139, 136, 135, 133, 131, 129, 123, 121, 121, 118, 117, 117, 114,
              112, 108, 104, 99, 99, 96, 94)
status <- c(rep(0, 5), rep(1, 67))
data <- reliability_data(x = cycles, status = status)

tbl_john <- estimate_cdf(data, methods = c("johnson", "kaplan"))

# Example 1: Probability Plot Weibull:
plot_weibull <- plot_prob(
  tbl_john,
  distribution = "weibull",
  title_main = "Weibull Analysis",
  title_x = "Cycles",
  title_y = "Probability of Failure in \%",
  title_trace = "Failed Items"
)

# Example 2: Probability Plot Lognormal:
plot_lognormal <- plot_prob(
  tbl_john,
  distribution = "lognormal",
  title_main = "Lognormal Analysis",
  title_x = "Cycles",
  title_y = "Probability of Failure in \%",
  title_trace = "Failed Items"
)

}
\references{
Meeker, William Q; Escobar, Luis A., Statistical methods for
  reliability data, New York: Wiley series in probability and statistics, 1998
}
