% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_prob.default}
\alias{plot_prob.default}
\title{Probability Plotting Method for Univariate Lifetime Distributions}
\usage{
\method{plot_prob}{default}(
  x,
  y,
  status,
  id = rep("XXXXXX", length(x)),
  distribution = c("weibull", "lognormal", "loglogistic", "normal", "logistic", "sev"),
  title_main = "Probability Plot",
  title_x = "Characteristic",
  title_y = "Unreliability",
  title_trace = "Sample",
  plot_method = c("plotly", "ggplot2"),
  ...
)
}
\arguments{
\item{x}{A numeric vector which consists of lifetime data. Lifetime
data could be every characteristic influencing the reliability of a product,
e.g. operating time (days/months in service), mileage (km, miles), load
cycles.}

\item{y}{A numeric vector which consists of estimated failure probabilities
regarding the lifetime data in \code{x}.}

\item{status}{A vector of binary data (0 or 1) indicating whether unit \emph{i}
is a right censored observation (= 0) or a failure (= 1).}

\item{id}{A character vector for the identification of every unit.}

\item{distribution}{Supposed distribution of the random variable.}

\item{title_main}{A character string which is assigned to the main title
of the plot.}

\item{title_x}{A character string which is assigned to the title of the
x axis.}

\item{title_y}{A character string which is assigned to the title of the
y axis.}

\item{title_trace}{A character string whis is assigned to the trace shown in
the legend.}

\item{plot_method}{Package, which is used for generating the plot output.}
}
\value{
Returns a plot object containing the probability plot.
}
\description{
This function is used to apply the graphical technique of probability
plotting.
}
\details{
If data was splitted by \code{\link{mixmod_em}} \code{\link{estimate_cdf}} with
method \code{"johnson"} is applied to subgroup-specific data. The
calculated plotting positions are shaped according to the determined split in
\code{\link{mixmod_em}}.

In \code{\link{mixmod_regression}} a maximum of three subgroups can be determined
and thus being plotted. The intention of this function is to give the
user a hint for the existence of a mixture model. An in-depth analysis should
be done afterwards.

The marker label for x and y are determined by the first word provided in the
argument \code{title_x} and \code{title_y} respectively, i.e. if
\code{title_x = "Mileage in km"} the x label of the marker is "Mileage".

The name of the legend entry is a combination of the \code{title_trace} and
the number of determined subgroups (if any). If \code{title_trace = "Group"}
and the data is split in two groups, the legend entries are "Group: 1"
and "Group: 2".
}
\examples{
# Alloy T7987 dataset taken from Meeker and Escobar(1998, p. 131)
cycles   <- c(300, 300, 300, 300, 300, 291, 274, 271, 269, 257, 256, 227, 226,
              224, 213, 211, 205, 203, 197, 196, 190, 189, 188, 187, 184, 180,
              180, 177, 176, 173, 172, 171, 170, 170, 169, 168, 168, 162, 159,
              159, 159, 159, 152, 152, 149, 149, 144, 143, 141, 141, 140, 139,
              139, 136, 135, 133, 131, 129, 123, 121, 121, 118, 117, 117, 114,
              112, 108, 104, 99, 99, 96, 94)
status <- c(rep(0, 5), rep(1, 67))

john <- estimate_cdf(
  x = cycles,
  status = status,
  methods = "johnson"
)

# Example 1: Probability Plot Weibull:
plot_weibull <- plot_prob(
  x = john$x,
  y = john$prob,
  status = john$status,
  id = john$id,
  distribution = "weibull",
  title_main = "Weibull Analysis",
  title_x = "Cycles",
  title_y = "Probability of Failure in \%",
  title_trace = "Failed Items"
)

# Example 2: Probability Plot Lognormal:
plot_lognormal <- plot_prob(
  x = john$x,
  y = john$prob,
  status = john$status,
  id = john$id,
  distribution = "lognormal",
  title_main = "Lognormal Analysis",
  title_x = "Cycles",
  title_y = "Probability of Failure in \%",
  title_trace = "Failed Items"
)

}
\references{
Meeker, William Q; Escobar, Luis A., Statistical methods for
reliability data, New York: Wiley series in probability and statistics, 1998
}
\seealso{
\code{\link{plot_prob}}
}
