% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\encoding{UTF-8}
\name{plot_prob}
\alias{plot_prob}
\title{Probability Plotting Method for Univariate Lifetime Distributions}
\usage{
plot_prob(x, ...)
}
\value{
Returns a plotly object containing the layout of the probability plot
  provided by \code{\link{plot_layout}} and the plotting positions.
}
\description{
This function is used to apply the graphical technique of probability
plotting.
}
\details{
The marker label for x and y are determined by the first word provided in the
argument \code{title_x} respective \code{title_y}, i.e. if
\code{title_x = "Mileage in km"} the x label of the marker is "Mileage".
}
\section{Methods (by class)}{

\describe{
  \item{\code{\link[=plot_prob.cdf_estimation]{cdf_estimation}}}{
    Preferred. Provide the output of \code{\link{estimate_cdf}} directly.
  }
  \item{\code{\link[=plot_prob.default]{default}}}{
    Provide \code{x}, \code{y}, \code{event} and \code{id} manually.
  }
}
}

\references{
Meeker, William Q; Escobar, Luis A., Statistical methods for
  reliability data, New York: Wiley series in probability and statistics, 1998
}
