% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_regression.R
\encoding{UTF-8}
\name{r_squared_profiling}
\alias{r_squared_profiling}
\title{R-Squared-Profile Function for Log-Location-Scale Distributions with Threshold}
\usage{
r_squared_profiling(x, ...)
}
\description{
This function evaluates the coefficient of determination with respect to a
given threshold parameter of a three-parametric lifetime distribution.
In terms of \emph{Median Rank Regression} this function can be optimized
(\code{\link{optim}}) to estimate the threshold parameter.
}
\references{
Meeker, William Q; Escobar, Luis A., Statistical methods for
  reliability data, New York: Wiley series in probability and statistics, 1998
}
