% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_regression.R
\name{rank_regression.default}
\alias{rank_regression.default}
\title{Rank Regression for Parametric Lifetime Distributions}
\usage{
\method{rank_regression}{default}(
  x,
  y,
  status,
  distribution = c("weibull", "lognormal", "loglogistic", "normal", "logistic", "sev",
    "weibull3", "lognormal3", "loglogistic3"),
  conf_level = 0.95,
  ...
)
}
\arguments{
\item{x}{A numeric vector which consists of lifetime data. Lifetime data
could be every characteristic influencing the reliability of a product,
e.g. operating time (days/months in service), mileage (km, miles), load
cycles.}

\item{y}{A numeric vector which consists of estimated failure probabilities
regarding the lifetime data in x.}

\item{status}{A vector of binary data (0 or 1) indicating whether a unit is
a right censored observation (= 0) or a failure (= 1).}

\item{distribution}{Supposed distribution of the random variable.}

\item{conf_level}{Confidence level of the interval. If \code{distribution} is
\code{"weibull"} this must be one of \code{0.9}, \code{0.95} or \code{0.99}.}
}
\value{
Returns a list with the following elements:
  \itemize{
    \item \code{coefficients} : If \code{distribution} is \code{"weibull"}, the
      estimated scale (\eqn{\eta}) and shape (\eqn{\beta}) parameters are provided
      (and additionally the threshold parameter \eqn{\gamma} if \code{distribution}
      is \code{"weibull3"}). For any other distribution, \code{coefficients} is
      identical to \code{loc_sc_params}.
    \item \code{confint} : Only included if \code{distribution} is \code{"weibull"}
      or \code{"weibull3"}. Approximated confidence intervals for \eqn{\eta} and
      \eqn{\beta} (and \eqn{\gamma} if \code{distribution} is \code{"weibull3"}).
    \item \code{loc_sc_params} : Estimated location-scale parameters. Threshold
      parameter is included if \code{distribution} is \code{"weibull3"},
      \code{"lognormal3"} or \code{"loglogistic3"}.
    \item \code{loc_sc_confint} : Confidence intervals for location-scale parameters.
      If distribution is \code{"lognormal3"} or \code{"loglogistic3"} a confidence
      interval for the threshold parameter is not computed.
    \item \code{loc_sc_varcov} : Provided, if \code{distribution} is not
      \code{"weibull"} or \code{"weibull3"}. Estimated heteroscedasticity-consistent
      variance-covariance matrix of the used location-scale distribution.
    \item \code{r_squared} : Coefficient of determination.
    \item \code{data} : A tibble returned from \code{\link{estimate_cdf}},
      excluding column \code{method}.
    \item \code{distribution} : Specified distribution.
  }
}
\description{
This method fits an \strong{x on y} regression to a linearized
two- or three-parameter cdf and is applicable for complete and (multiple) right
censored data. The parameters are estimated in the frequently used (log-) location-scale
parameterization. For the Weibull, estimates are transformed such that
they are in line with the parameterization provided by the \emph{stats} package
(see \link[stats]{Weibull}).
}
\details{
If \code{distribution} is \code{"weibull"} or \code{"weibull3"}, the approximated
confidence intervals for the parameters can only be estimated on the following
confidence levels (see 'References' \emph{(Mock, 1995)}):
\itemize{
  \item \code{conf_level} = 0.90,
  \item \code{conf_level} = 0.95,
  \item \code{conf_level} = 0.99.
}

If the distribution is not the Weibull, the confidence intervals of the
parameters are computed on the basis of a heteroscedasticity-consistent
covariance matrix. Here it should be said that there is no statistical foundation
to determine the standard errors of the parameters using \emph{Least Squares}
in context of \emph{Rank Regression}. For an accepted statistical method use MLE
(\code{\link{ml_estimation}}).
}
\examples{
# Vectors:
obs <- seq(10000, 100000, 10000)
status_1 <- c(0, 1, 1, 0, 0, 0, 1, 0, 1, 0)

cycles <- alloy$cycles
status_2 <- alloy$status

# Example 1 - Fitting a two-parametric weibull distribution:
tbl_john <- estimate_cdf(
  x = obs,
  status = status_1,
  methods = "johnson"
)

rr <- rank_regression(
  x = tbl_john$x,
  y = tbl_john$prob,
  status = tbl_john$status,
  distribution = "weibull",
  conf_level = .90
)

# Example 2: Fitting a three-parametric lognormal distribution:
tbl_kaplan <- estimate_cdf(
  x = cycles,
  status = status_2,
  methods = "kaplan"
)

rr_2 <- rank_regression(
  x = tbl_kaplan$x,
  y = tbl_kaplan$prob,
  status = tbl_kaplan$status,
  distribution = "lognormal3"
)

}
\references{
\itemize{
    \item Mock, R., Methoden zur Datenhandhabung in
  Zuverlässigkeitsanalysen, vdf Hochschulverlag AG an der ETH Zürich, 1995
    \item Meeker, William Q; Escobar, Luis A., Statistical methods for
  reliability data, New York: Wiley series in probability and statistics, 1998}
}
\seealso{
\code{\link{rank_regression}}
}
