% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{voltage}
\alias{voltage}
\title{High Voltage Stress Test for the Dielectric Insulation of Generator armature bars}
\format{
A tibble with 58 rows and 3 variables:
\describe{
  \item{hours}{Observed hours.}
  \item{failure_mode}{
    One of two failure modes (\code{D} and \code{E})
    or \code{censored} if no failure occured.
  }
  \item{status}{
    If \code{failure_mode} is either \code{D} or \code{E}
    this is \code{1} else \code{0}.
  }
}
}
\source{
Doganaksoy, N.; Hahn, G.; Meeker, W. Q., Reliability Analysis by
Failure Mode, Quality Progress, 35(6), 47-52, 2002
}
\usage{
voltage
}
\description{
A sample of 58 segments of bars were subjected to a high voltage stress test.
Two failure modes occured, Mode D (degradation failure) and Mode E (early failure).
}
\keyword{datasets}
