% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculate_ranks}
\alias{calculate_ranks}
\title{Computation of Johnson Ranks}
\usage{
calculate_ranks(f, n_out, n)
}
\arguments{
\item{f}{a numeric vector indicating the number of failed units for a
specific realization of the lifetime characteristic.}

\item{n_out}{a numeric vector indicating the number of failed and censored
units that have a shorter realization of lifetime characteristic as unit
\emph{i}.}

\item{n}{an integer value indicating the sample size.}
}
\value{
A numeric vector containing the computed Johnson ranks.
}
\description{
This function calculates the Johnson ranks which are used to estimate the
failure probabilities in case of (multiple) right censored data.
}
\keyword{internal}
