% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\encoding{UTF-8}
\name{confint_betabinom.default}
\alias{confint_betabinom.default}
\title{Beta Binomial Confidence Bounds for Quantiles and/or Probabilities}
\usage{
\method{confint_betabinom}{default}(
  x,
  status,
  loc_sc_params,
  distribution = c("weibull", "lognormal", "loglogistic", "normal", "logistic", "sev",
    "weibull3", "lognormal3", "loglogistic3"),
  b_lives = c(0.01, 0.1, 0.5),
  bounds = c("two_sided", "lower", "upper"),
  conf_level = 0.95,
  direction = c("y", "x")
)
}
\arguments{
\item{x}{A numeric vector which consists of lifetime data. \code{x} is used to
specify the range of confidence region(s).}

\item{status}{A vector of binary data (0 or 1) indicating whether unit \emph{i}
is a right censored observation (= 0) or a failure (= 1).}

\item{loc_sc_params}{A (named) numeric vector of estimated location
and scale parameters for a specified distribution. The order of
elements is important. First entry needs to be the location
parameter \eqn{\mu} and the second element needs to be the scale
parameter \eqn{\sigma}. If a three-parametric model is used the third element
is the threshold parameter \eqn{\gamma}.}

\item{distribution}{Supposed distribution of the random variable.}

\item{bounds}{A character string specifying the interval(s) which has/have to
be computed. Must be one of "two_sided" (default), "lower" or "upper".}

\item{conf_level}{Confidence level of the interval. The default value is
\code{conf_level = 0.95}.}

\item{direction}{A character string specifying the direction of the computed
interval(s). Must be either "y" (failure probabilities) or "x" (quantiles).}
}
\value{
A tibble containing the lifetime characteristic, interpolated
ranks as a function of probabilities, the probabilities which are used to
compute the ranks and computed values for the specified confidence bound(s).
}
\description{
This non-parametric approach calculates confidence bounds for quantiles and/or
failure probabilities using a procedure that is similar to that used in
calculating median ranks. The location-scale (and threshold) parameters estimated
by rank regression are needed.
}
\examples{
# Example 1: Beta-Binomial Confidence Bounds for two-parameter Weibull:
obs   <- seq(10000, 100000, 10000)
status <- c(0, 1, 1, 0, 0, 0, 1, 0, 1, 0)

tbl <- reliability_data(x = obs, status = status)

tbl_john <- estimate_cdf(tbl, methods = "johnson")

mrr <- rank_regression(
  x = tbl_john$x,
  y = tbl_john$prob,
  status = tbl_john$status,
  distribution = "weibull",
  conf_level = .95
)

conf_betabin <- confint_betabinom(
  x = tbl_john$x,
  status = tbl_john$status,
  loc_sc_params = mrr$loc_sc_params,
  distribution = "weibull"
  bounds = "two_sided",
  conf_level = 0.95,
  direction = "y"
)

# Example 2: Beta-Binomial Confidence Bounds for three-parameter Weibull:
# Alloy T7987 dataset taken from Meeker and Escobar(1998, p. 131)
tbl_2 <- reliability_data(shock, x = cycles, status = status)

tbl_john_2 <- estimate_cdf(tbl_2, methods = "johnson")

mrr_weib3 <- rank_regression(
  x = tbl_john_2$x,
  y = tbl_john_2$prob,
  status = tbl_john_2$status,
  distribution = "weibull3",
  conf_level = .95
)

conf_betabin_weib3 <- confint_betabinom(
  x = tbl_john_2$x,
  status = tbl_john_2$status,
  loc_sc_params = mrr_weib3$loc_sc_params,
  distribution = "weibull3",
  bounds = "two_sided",
  conf_level = 0.95,
  direction = "y"
)

}
\references{
Meeker, William Q; Escobar, Luis A., Statistical methods for
  reliability data, New York: Wiley series in probability and statistics, 1998
}
