% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{confint_betabinom}
\alias{confint_betabinom}
\title{Beta Binomial Confidence Bounds for Quantiles and/or Probabilities}
\usage{
confint_betabinom(x, ...)
}
\value{
A tibble containing the lifetime characteristic, interpolated
ranks as a function of probabilities, the probabilities which are used to
compute the ranks and computed values for the specified confidence bound(s).
}
\description{
This non-parametric approach calculates confidence bounds for quantiles and/or
failure probabilities using a procedure that is similar to that used in
calculating median ranks. The location-scale (and threshold) parameters estimated
by rank regression are needed.
}
\section{Methods (by class)}{

\describe{
  \item{\code{\link[=confint_betabinom.model_estimation]{model_estimation}}}{
    Preferred. Provide the output of \code{\link{rank_regression}}.
   }
  \item{\code{\link[=confint_betabinom.default]{default}}}{
    Provide \code{x}, \code{event}, \code{loc_sc_params} and \code{distribution}
    manually.
  }
}
}

