% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{confint_fisher}
\alias{confint_fisher}
\title{Fisher Confidence Bounds for Quantiles and/or Probabilities}
\usage{
confint_fisher(x, ...)
}
\value{
A tibble containing the lifetime characteristic, the
  probabilities, estimated standard errors by the delta method and computed
  values for the specified confidence bound(s).
}
\description{
This method computes normal-approximation confidence intervals for quantiles
and/or failure probabilities using the \code{\link{delta_method}}. The
required (log-)location-scale (and threshold) parameters and variance-covariance matrix
of these need to be estimated by Maximum Likelihood.
}
\section{Methods (by class)}{

\describe{
  \item{\code{\link[=confint_fisher.parameter_estimation]{parameter_estimation}}}{
    Preferred. Provide the output of \code{\link{ml_estimation}}.
  }
  \item{\code{\link[=confint_fisher.default]{default}}}{
    Provide \code{x}, \code{event}, \code{loc_sc_params}, \code{loc_sc_varcov}
    and \code{distribution} manually.
  }
}
}

