% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_estimators.R
\name{estimate_cdf}
\alias{estimate_cdf}
\title{Estimation of Failure Probabilities}
\usage{
estimate_cdf(data, methods, options = list())
}
\arguments{
\item{data}{A tibble returned by \link{reliability_data}.}

\item{methods}{One or multiple methods of \code{"mr"}, \code{"johnson"},
\code{"kaplan"} or \code{"nelson"} used for estimating the failure
probabilities. See 'Details'.}

\item{options}{A list of named options. See 'Options'.}
}
\value{
A tibble containing the following columns:
\itemize{
  \item \code{id}: Identification.
  \item \code{characteristic}: Lifetime characteristic.
  \item \code{status}: Binary data (0 or 1) indicating whether a unit is a
    right censored observation (=0) or a failure (=1).
  \item \code{rank}: Adjusted rank. Applicable for method \code{"mr"}
    and \code{"johnson"}, otherwise \code{NA}.
  \item \code{prob}: Estimated failure probability.
  \item \code{method}: Method, which was used for estimating the failure
    probability.
}
}
\description{
Estimate failure probabilities for uncensored or (multiple) right-censored
data.
}
\details{
One or multiple of the following methods can be used as the \code{methods}
argument:
\itemize{
  \item \code{"mr"}: This non-parametric approach (\emph{Median Ranks}) is
    used to estimate the failure probabilities in terms of complete data.
    Tied observations can be handled in three ways, "max", "min" and "average"
    (usage is listed in 'Options')).
    Two methods are available to estimate the cumulative distribution
    function \emph{F(t)} (See 'Options'):
    \itemize{
      \item "benard"; Benard's approximation for Median Ranks.
      \item "invbeta"; Exact Median Ranks using the inverse beta distribution.
    }
  \item \code{"johnson"}: This non-parametric approach is used to estimate
    the failure probabilities in terms of uncensored or (multiple) right
    censored data. Compared to complete data the correction is done by
    calculating adjusted ranks which takes non-defective units into account.
  \item \code{"kaplan"}: Whereas the non-parametric Kaplan-Meier estimator is
    used to estimate the survival function _S(t)_ in terms of (multiple)
    right censored data, the complement is an estimate of the cumulative
    distribution function _F(t)_. One modification is made in contrast
    to the orginial Kaplan-Meier estimator (based on
    _NIST/SEMATECH e-Handbook of Statistical Methods_, 8.2.1.5.): If the
    last unit (unit with highest observed lifetime) is a defective unit, the
    estimator is adjusted in such a way that the survival estimate for this
    unit is not _zero_ and therefore the estimate for the failure
    probability is not equal to _one_. Otherwise the estimate in this
    context would be too pessimisitc.
    Since the failure probability estimation in this function is not based on
    _Median Ranks_, the Betabinomial confidence intervals _cannot_ be
    calculated on the basis of Kaplan-Meier failure probabilities.
  \item \code{"nelson"}: This non-parametric approach estimates the cumulative
    hazard rate in terms of (multiple) right censored data. By equating the
    definition of the hazard rate with the hazard rate according to
    Nelson-Aalen one can calculate the failure probabilities.
    Since the failure probability estimation in this function is not based on
    _Median Ranks_, the Betabinomial confidence intervals cannot be calculated
    on the basis of Nelson-Aalen failure probabilities.
}
}
\section{Options}{


\itemize{
  \item \code{mr_method}: \code{"benard"} (default) or \code{"invbeta"}.
  \item \code{mr_ties.method}: \code{c("max", "min", "average")}, default is "max".
}
}

