% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_estimators.R
\name{estimate_cdf}
\alias{estimate_cdf}
\title{Estimation of Failure Probabilities}
\usage{
estimate_cdf(reliability_tbl, methods, options = list())
}
\arguments{
\item{reliability_tbl}{A tibble returned by \link{reliability_data}.}

\item{methods}{Character vector of methods used for estimating the failure
probabilities. See 'Details'.}

\item{options}{A list of named options passed to \code{<method>_method}. For
now there is just the option "method" with \code{method = "mr"}. See
\code{\link{mr_method}}.}
}
\value{
A tibble containing the following columns:
\itemize{
  \item \code{id}: Identification.
  \item \code{characteristic}: Lifetime characteristic.
  \item \code{status}: Binary data (0 or 1) indicating whether a unit is a
  right censored observation (=0) or a failure (=1).
  \item \code{rank}: Adjusted rank. Applicable for method \code{"mr"}
  and \code{"johnson"}, otherwise \code{NA}.
  \item \code{prob}: Estimated failure probability.
  \item \code{method}: Method for estimating the failure probability.
}
}
\description{
Estimate failure probabilities for uncensored or (multiple) right-censored
data.
}
\details{
Depending on the \code{methods} argument one or more of the following functions
are called:
\itemize{
  \item \code{"mr"}: \code{\link{mr_method}}. Works only for uncensored data.
  \item \code{"johnson"}: \code{\link{johnson_method}}.
  \item \code{"kaplan"}: \code{\link{kaplan_method}}.
  \item \code{"nelson"}: \code{\link{nelson_method}}.
}
}
