% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_estimation.R
\encoding{UTF-8}
\name{ml_estimation}
\alias{ml_estimation}
\title{ML Estimation for Parametric Lifetime Distributions}
\usage{
ml_estimation(x, ...)
}
\value{
Returns a list with the following components:
  \itemize{
  \item \code{coefficients} : Provided, if \code{distribution} is \code{"weibull"}.
    \eqn{\eta} is the estimated scale and \eqn{\beta} the estimated shape parameter.
  \item \code{confint} : Provided, if \code{distribution} is \code{"weibull"}.
    Confidence interval for \eqn{\eta} and \eqn{\beta}.
  \item \code{loc_sc_params} : Estimated location-scale parameters.
  \item \code{loc_sc_confint} : Confidence interval for location-scale parameters.
  \item \code{loc_sc_varcov} : Estimated Variance-Covariance matrix of the used
    location-scale distribution.
  \item \code{logL} : The log-likelihood value.
  \item \code{aic} : Akaike Information Criterion.
  \item \code{bic} : Bayesian Information Criterion.}
}
\description{
This method estimates the parameters and calculates normal approximation confidence
intervals for a two- or three-parametric lifetime distribution in the frequently used
(log-) location-scale parameterization. \code{ml_estimation} uses the
\code{\link{Lifedata.MLE}} function which is defined in the
\emph{SPREDA} package.
For the Weibull the estimates are transformed such that they are in line with
the parameterization provided by the \emph{stats} package like
\code{\link{pweibull}}. The method is applicable for complete and (multiple)
right censored data.
}
\section{Methods (by class)}{

\describe{
  \item{\code{\link[=ml_estimation.reliability_data]{reliability_data}}}{
    Preferred. Provide the output of \code{\link{reliability_data}} directly.
  }
  \item{\code{\link[=ml_estimation.default]{default}}}{
    Provide \code{x} and \code{status} manually.
  }
}
}

\references{
Meeker, William Q; Escobar, Luis A., Statistical methods for
  reliability data, New York: Wiley series in probability and statistics, 1998
}
