% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_estimators.R
\name{mr_method}
\alias{mr_method}
\title{Estimation of Failure Probabilities using Median Ranks}
\usage{
mr_method(
  x,
  status = rep(1, length(x)),
  id = rep("XXXXXX", length(x)),
  method = c("benard", "invbeta"),
  ties.method = c("max", "min", "average")
)
}
\arguments{
\item{x}{A numeric vector which consists of lifetime data. Lifetime
data could be every characteristic influencing the reliability of a product,
e.g. operating time (days/months in service), mileage (km, miles), load
cycles.}

\item{status}{A vector of ones indicating that every unit \emph{i} has failed.}

\item{id}{A character vector for the identification of every unit.}

\item{method}{Method for the estimation of the cdf. Can be "benard" (default)
or "invbeta".}

\item{ties.method}{A character string specifying how ties are treated, default is "max".}
}
\value{
A tibble containing id, lifetime characteristic, status of the
  unit, the rank and the estimated failure probability.
}
\description{
\lifecycle{soft-deprecated}

The functionality of this function is incorporated into \code{\link{estimate_cdf}}.

This non-parametric approach (\emph{Median Ranks}) is used to estimate the
failure probabilities in terms of complete data. Two methods are available to
estimate the cumulative distribution function \emph{F(t)}:
\itemize{
  \item "benard"; Benard's approximation for Median Ranks
  \item "invbeta"; Exact Median Ranks using the inverse beta distribution
}
}
\examples{
# Example 1
obs   <- seq(10000, 100000, 10000)
state <- rep(1, length(obs))
uic   <- c("3435", "1203", "958X", "XX71", "abcd", "tz46",
           "fl29", "AX23", "Uy12", "kl1a")

tbl_mr <- mr_method(x = obs, status = state, id = uic,
                   method = "benard")

# Example 2
tbl_mr_invbeta <- mr_method(x = obs, status = state, id = uic,
                           method = "invbeta")

}
