% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\encoding{UTF-8}
\name{plot_conf}
\alias{plot_conf}
\title{Add Confidence Region(s) for Quantiles or Probabilities}
\usage{
plot_conf(p_obj, x, ...)
}
\value{
Returns a plot object containing the probability plot with
  plotting positions, the estimated regression line and the estimated
  confidence region(s).
}
\description{
This function is used to add estimated confidence region(s) to an existing
probability plot which also includes the estimated regression line.
}
\section{Methods (by class)}{

\describe{
  \item{\code{\link[=plot_conf.confint]{confint}}}{
    Preferred. Provide the output of either \code{\link{confint_betabinom}}
    or \code{\link{confint_fisher}} directly.
  }
  \item{\code{\link[=plot_conf.default]{default}}}{
    Provide \code{x}, \code{y}, \code{distribution} and \code{direction}
    manually.
  }
}
}

\references{
Meeker, William Q; Escobar, Luis A., Statistical methods for
  reliability data, New York: Wiley series in probability and statistics, 1998
}
