% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_mod.default}
\alias{plot_mod.default}
\title{Add Estimated Population Line to a Probability Plot}
\usage{
\method{plot_mod}{default}(
  p_obj,
  x,
  loc_sc_params,
  distribution = c("weibull", "lognormal", "loglogistic", "normal", "logistic", "sev",
    "weibull3", "lognormal3", "loglogistic3"),
  title_trace = "Fit",
  ...
)
}
\arguments{
\item{p_obj}{A plot object returned from \code{\link{plot_prob}}.}

\item{x}{A numeric vector containing the x-coordinates of the regression line.}

\item{loc_sc_params}{A (named) numeric vector of estimated location
and scale parameters for a specified distribution. The order of
elements is important. First entry needs to be the location
parameter \eqn{\mu} and the second element needs to be the scale
parameter \eqn{\sigma}. If a three-parametric model is used the third element
is the threshold parameter \eqn{\gamma}.}

\item{distribution}{Supposed distribution of the random variable.}

\item{title_trace}{A character string whis is assigned to the trace shown in
the legend.}

\item{...}{Further arguments passed to or from other methods. Currently not used.}
}
\value{
Returns a plot object containing the probability plot with
plotting positions and the estimated regression line(s).
}
\description{
This function adds one or multiple estimated regression lines to an existing
probability plot (\code{\link{plot_prob}}). Depending on the output of the
functions \code{\link{rank_regression}}, \code{ml_estimation},
\code{\link{mixmod_regression}} or \code{\link{mixmod_em}} one or
multiple lines are plotted.
}
\examples{
# Alloy T7987 dataset taken from Meeker and Escobar(1998, p. 131)
cycles   <- c(300, 300, 300, 300, 300, 291, 274, 271, 269, 257, 256, 227, 226,
              224, 213, 211, 205, 203, 197, 196, 190, 189, 188, 187, 184, 180,
              180, 177, 176, 173, 172, 171, 170, 170, 169, 168, 168, 162, 159,
              159, 159, 159, 152, 152, 149, 149, 144, 143, 141, 141, 140, 139,
              139, 136, 135, 133, 131, 129, 123, 121, 121, 118, 117, 117, 114,
              112, 108, 104, 99, 99, 96, 94)
status <- c(rep(0, 5), rep(1, 67))

tbl_john <- estimate_cdf(x = cycles, status = status, method = "johnson")

# Example 1: Probability Plot and Regression Line Three-Parameter-Weibull:
plot_weibull <- plot_prob(
  x = tbl_john$x,
  y = tbl_john$prob,
  status = tbl_john$status,
  id = tbl_john$id,
  distribution = "weibull",
  title_main = "Three-Parametric Weibull",
  title_x = "Cycles",
  title_y = "Probability of Failure in \%",
  title_trace = "Failed Items"
)

rr <- rank_regression(
  x = tbl_john$x,
  y = tbl_john$prob,
  status = tbl_john$status,
  distribution = "weibull3",
  conf_level = 0.9
)

plot_reg_weibull <- plot_mod(
  p_obj = plot_weibull,
  x = tbl_john$x,
  loc_sc_params = rr$loc_sc_params,
  distribution = "weibull3",
  title_trace = "Estimated Weibull CDF"
)



# Example 2: Probability Plot and Regression Line Three-Parameter-Lognormal:
plot_lognormal <- plot_prob(
  x = tbl_john$x,
  y = tbl_john$prob,
  status = tbl_john$status,
  id = tbl_john$id,
  distribution = "lognormal",
  title_main = "Three-Parametric Lognormal",
  title_x = "Cycles",
  title_y = "Probability of Failure in \%",
  title_trace = "Failed Items"
)

rr_ln <- rank_regression(
  x = tbl_john$x,
  y = tbl_john$prob,
  status = tbl_john$status,
  distribution = "lognormal3",
  conf_level = 0.9
)

plot_reg_lognormal <- plot_mod(
  p_obj = plot_lognormal,
  x = tbl_john$x,
  loc_sc_params = rr_ln$loc_sc_params,
  distribution = "lognormal3",
  title_trace = "Estimated Lognormal CDF"
)

}
\references{
Meeker, William Q; Escobar, Luis A., Statistical methods for
  reliability data, New York: Wiley series in probability and statistics, 1998
}
\seealso{
\code{\link{plot_mod}}
}
