% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\encoding{UTF-8}
\name{plot_mod}
\alias{plot_mod}
\title{Add Estimated Population Line(s) to a Probability Plot}
\usage{
plot_mod(p_obj, x, ...)
}
\value{
Returns a plotly object containing the probability plot with
  plotting positions and the estimated regression line.
}
\description{
This function adds one or multiple estimated regression lines to an existing
probability plot (\code{\link{plot_prob}}). Depending on the output of the
functions \code{\link{rank_regression}}, \code{ml_estimation},
\code{\link{mixmod_regression}} or \code{\link{mixmod_em}} one or
multiple lines are plotted.
}
\details{
The name of the legend entry is a combination of the \code{title_trace} and
the number of determined subgroups from \code{\link{mixmod_regression}} or
\code{\link{mixmod_em}}. If \code{title_trace = "Line"} and the
data could be splitted in two groups, the legend entries would be "Line: 1"
and "Line: 2".
}
\section{Methods (by class)}{

\describe{
  \item{\code{\link[=plot_mod.model_estimation]{model_estimation(_list)}}}{
    Preferred. Provide the output of either \code{\link{ml_estimation}} or
    \code{\link{rank_regression}} directly.
  }
  \item{\code{\link[=plot_mod.default]{default}}}{
    Provide \code{x}, \code{loc_sc_params} and \code{distribution} manually.
  }
}
}

\references{
Meeker, William Q; Escobar, Luis A., Statistical methods for
  reliability data, New York: Wiley series in probability and statistics, 1998
}
