% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\encoding{UTF-8}
\name{plot_mod}
\alias{plot_mod}
\alias{plot_mod.model_estimation}
\alias{plot_mod.model_estimation_list}
\alias{plot_mod.mixmod_regression}
\alias{plot_mod.mixmod_regression_list}
\alias{plot_mod.mixmod_em}
\title{Add Estimated Population Line(s) to a Probability Plot}
\usage{
plot_mod(p_obj, x, ...)

## S3 method for classes 'model_estimation', 'model_estimation_list',
## 'mixmod_regression', 'mixmod_regression_list', 'mixmod_em'
plot_mod(p_obj, x, title_trace = "Fit")
}
\arguments{
\item{p_obj}{A plot object returned from \code{\link{plot_prob}}.}

\item{x}{An object of aforementioned classes. See 'S3 methods'.}

\item{...}{Further arguments passed to or from other methods. Currently not used.}

\item{title_trace}{A character string which is assigned to the trace shown in
the legend.}
}
\value{
Returns a plot object containing the probability plot with
  plotting positions and the estimated regression line(s).
}
\description{
This function adds one or multiple estimated regression lines to an existing
probability plot (\code{\link{plot_prob}}). Depending on the output of the
functions \code{\link{rank_regression}}, \code{ml_estimation},
\code{\link{mixmod_regression}} or \code{\link{mixmod_em}} one or
multiple lines are plotted.
}
\details{
The name of the legend entry is a combination of the \code{title_trace} and
the number of determined subgroups from \code{\link{mixmod_regression}} or
\code{\link{mixmod_em}}. If \code{title_trace = "Line"} and the
data could be splitted in two groups, the legend entries would be "Line: 1"
and "Line: 2".
}
\section{S3 methods}{
\tabular{lll}{
   \strong{Class} \tab \strong{Returned from} \tab \strong{Condition} \cr
   \code{model_estimation} \tab \code{\link{rank_regression}} \tab One method in \code{\link{estimate_cdf}}. \cr
   \code{model_estimation} \tab \code{\link{mixmod_regression}} \tab One method in \code{\link{estimate_cdf}}. No subgroups determined. \cr
   \code{model_estimation} \tab \code{\link{ml_estimation}} \tab - \cr
   \code{model_estimation_list} \tab \code{\link{rank_regression}} \tab Multiple methods in \code{\link{estimate_cdf}}. \cr
   \code{mixmod_regression} \tab \code{\link{mixmod_regression}} \tab One method in \code{\link{estimate_cdf}}. \cr
   \code{mixmod_regression_list} \tab \code{\link{mixmod_regression}} \tab Multiple methods in \code{\link{estimate_cdf}}. \cr
   \code{mixmod_em} \tab \code{\link{mixmod_em}} \tab - \cr
}
}

\examples{
# Reliability data:
data <- reliability_data(
  data = alloy,
  x = cycles,
  status = status
)

# Probability estimation:
prob_tbl <- estimate_cdf(data, methods = c("johnson", "kaplan"))

# Example 1 - Probability Plot and Regression Line Three-Parameter-Weibull:
plot_weibull <- plot_prob(
  prob_tbl,
  distribution = "weibull"
)

rr_weibull <- rank_regression(
  prob_tbl,
  distribution = "weibull3"
)

plot_reg_weibull <- plot_mod(
  p_obj = plot_weibull,
  x = rr_weibull
)

# Example 2 - Probability Plot and Regression Line Three-Parameter-Lognormal:
plot_lognormal <- plot_prob(
  prob_tbl,
  distribution = "lognormal"
)

rr_lognormal <- rank_regression(
  prob_tbl,
  distribution = "lognormal3"
)

plot_reg_lognormal <- plot_mod(
  p_obj = plot_lognormal,
  x = rr_lognormal
)

}
\references{
Meeker, William Q; Escobar, Luis A., Statistical methods for
  reliability data, New York: Wiley series in probability and statistics, 1998
}
