% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\encoding{UTF-8}
\name{plot_mod}
\alias{plot_mod}
\title{Adding an Estimated Population Line to a Probability Plot}
\usage{
plot_mod(p_obj, x, ...)
}
\value{
Returns a plotly object containing the probability plot with
  plotting positions and the estimated regression line.
}
\description{
This function adds a regression line to an existing probability plot using a
model estimated by \code{\link{rank_regression}} or \code{\link{ml_estimation}}.
}
\section{Methods (by class)}{

\describe{
  \item{\code{\link[=plot_mod.model_estimation]{model_estimation}}}{
    Preferred. Provide the output of either \code{\link{ml_estimation}} or
    \code{\link{rank_regression}}.
  }
  \item{\code{\link[=plot_mod.default]{default}}}{
    Provide \code{x}, \code{loc_sc_params} and \code{distribution} manually.
  }
}
}

\references{
Meeker, William Q; Escobar, Luis A., Statistical methods for
  reliability data, New York: Wiley series in probability and statistics, 1998
}
