% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_pop}
\alias{plot_pop}
\title{Add Population Line(s) to an Existing Grid}
\usage{
plot_pop(
  p_obj = NULL,
  x,
  loc_sc_params_tbl,
  distribution = c("weibull", "lognormal", "loglogistic"),
  tol = 1e-06,
  title_trace = "Population",
  plot_method = c("plotly", "ggplot2")
)
}
\arguments{
\item{p_obj}{A plot object to which the population lines are added or
\code{NULL}. If \code{NULL} the population lines are added to an empty grid.}

\item{x}{A numeric vector of length two or greater used for the x coordinates
of the population line. If \code{length(x) == 2} a sequence of length 200
between \code{x[1]} and \code{x[2]} is created. This sequence is equidistant
with respect to the scale of the x axis. If \code{length(x) > 2} the elements
of \code{x} are the x coordinates of the population line.}

\item{loc_sc_params_tbl}{A tibble. See 'Details'.}

\item{tol}{The failure probability is restricted to the interval
\eqn{[tol, 1 - tol]}. The default value is in accordance with the decimal
places shown in the hover for \code{plot_method = "plotly"}.}

\item{plot_method}{Plot package, which produces the visual output. Only
used with \code{p_obj = NULL}, otherwise \code{p_obj} is used to determine
the plot method.}
}
\value{
A plot object which contains the linearized CDF(s).
}
\description{
This function adds one or multiple linearized CDFs to an existing plot grid.
}
\details{
\code{loc_sc_params_tbl} is a tibble with the following columns:
  \itemize{
    \item \code{loc}: Location parameter \eqn{\mu}.
    \item \code{sc}: Scale parameter \eqn{\sigma}.
  }
  If only one line should be displayed, a numeric vector of length two is
  also supported \code{c(loc, sc)}.
}
\examples{
x <- rweibull(n = 100, shape = 1, scale = 20000)

grid_weibull <- plot_layout(
  x = x,
  distribution = "weibull",
  title_main = "Weibull Analysis",
  title_x = "Time to Failure",
  title_y = "Failure Probability"
)
# Example 1: One line
pop_weibull <- plot_pop(
  p_obj = grid_weibull,
  x = range(x),
  loc_sc_params_tbl = c(log(20000), 1),
  distribution = "weibull",
  title_trace = "Population"
)

# Example 2: Multiple lines
pop_weibull2 <- plot_pop(
  p_obj = NULL,
  x = x,
  loc_sc_params_tbl = tibble(
    loc = c(log(20000), log(20000), log(20000)),
    sc = c(1, 1.5, 2)
    ),
  distribution = "weibull",
  title_trace = "Population",
  plot_method = "ggplot2"
)

}
