% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\encoding{UTF-8}
\name{plot_prob_mix}
\alias{plot_prob_mix}
\title{Probability Plot for Separated Mixture Models}
\usage{
plot_prob_mix(x, ...)
}
\description{
This function is used to apply the graphical technique of probability
plotting to univariate mixture models that were separated with functions
\code{\link{mixmod_regression}} or \code{\link{mixmod_em}}.
}
\details{
Depending on the separation method the function \code{\link{johnson_method}}
is called in various ways. If \code{mixmod_regression} is used, \code{johnson_method}
is applied to all data. If data was splitted by \code{mixmod_em} the function
\code{johnson_method} is applied to subgroup-specific data. The calculated plotting
positions are colored regarding the obtained split of the used splitting function.
If \code{mix_output = NULL} \code{johnson_method} is applied to all data, too.
The obtained plot is then equal to \code{\link{plot_prob}}. See \strong{Examples}
for all three cases.

In \code{\link{mixmod_regression}} a maximum of three subgroups can be determined
and thus being plotted. The intention of this function is to give the
user a hint for the existence of a mixture model. An in-depth analysis should
be done afterwards.

The marker label for x is determined by the first word provided in the
argument \code{title_x}, i.e. if \code{title_x = "Mileage in km"} the x label
of the marker is "Mileage".

The marker label for y is determined by the string provided in the
argument \code{title_y}, i.e. if \code{title_y = "Probability in percent"} the y
label of the marker is "Probability".

The name of the legend entry is a combination of the \code{title_trace} and the
number of determined subgroups. If \code{title_trace = "Group"} and the data
could be splitted in two groups, the legend entries would be "Group 1" and "Group 2".
}
\references{
Doganaksoy, N.; Hahn, G.; Meeker, W. Q., Reliability Analysis by
  Failure Mode, Quality Progress, 35(6), 47-52, 2002
}
