% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{predict_prob}
\alias{predict_prob}
\title{Estimation of Failure Probabilities for Parametric Lifetime Distributions}
\usage{
predict_prob(
  q,
  loc_sc_params,
  distribution = c("weibull", "lognormal", "loglogistic", "normal", "logistic", "sev",
    "weibull3", "lognormal3", "loglogistic3")
)
}
\arguments{
\item{q}{A numeric vector which consists of lifetime data.}

\item{loc_sc_params}{A (named) numeric vector of estimated location
and scale parameters for a specified distribution. The order of
elements is important. First entry needs to be the location
parameter \eqn{\mu} and the second element needs to be the scale
parameter \eqn{\sigma}. If a three-parametric model is used the third element
is the threshold parameter \eqn{\gamma}.}

\item{distribution}{Supposed distribution of the random variable.}
}
\value{
A vector containing the estimated failure probabilities for a given
  set of quantiles and estimated parameters.
}
\description{
This function estimates the failure probabilities for a given set of estimated
location-scale (and threshold) parameters and specified quantiles.
}
\examples{
# Example 1: Predicted probabilities for two-parameter Weibull:
probs <- predict_prob(
  q = c(15, 48, 124),
  loc_sc_params = c(5, 0.5),
  distribution = "weibull"
)

# Example 2: Predicted probabilities for three-parameter Weibull:
probs_weib3 <- predict_prob(
  q = c(25, 58, 134),
  loc_sc_params = c(5, 0.5, 10),
  distribution = "weibull3"
)
}
