% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_regression.R
\encoding{UTF-8}
\name{rank_regression}
\alias{rank_regression}
\title{Rank Regression for Parametric Lifetime Distributions}
\usage{
rank_regression(x, ...)
}
\value{
Returns a list with the following components:
  \itemize{
  \item \code{coefficients} : Provided, if \code{distribution} is \code{"weibull"}
    or \code{"weibull3"}. \eqn{\eta} is the estimated scale and \eqn{\beta}
    the estimated shape parameter. The estimated threshold parameter \eqn{\gamma}
    is available if the three-parametric weibull is used.
  \item \code{confint} : Provided, if \code{distribution} is \code{"weibull"}
    or \code{"weibull3"}. Confidence intervals for \eqn{\eta} and \eqn{\beta}
    (and \eqn{\gamma} if the three-parametric weibull is used).
  \item \code{loc_sc_params} : Estimated location-scale parameters.
    Threshold parameter is provided for \code{"weibull3"}, \code{"lognormal3"}
    and \code{"loglogistic3"}.
  \item \code{loc_sc_confint} : Confidence intervals for location-scale parameters.
    If distribution is \code{"lognormal3"} or \code{"loglogistic3"} a confidence
    interval for the threshold is not computed.
  \item \code{loc_sc_varcov} : Provided, if \code{distribution} is not
    \code{"weibull"} or \code{"weibull3"}. Estimated heteroscedasticity-consistent
    Variance-Covariance matrix of the used location-scale distribution.
  \item \code{r_squared} : Coefficient of determination.}
}
\description{
This method fits an \strong{x on y} regression to a linearized
two- or three-parameter cdf and is applicable for complete and (multiple) right
censored data. The parameters are estimated in the frequently used (log-) location-scale
parameterization. For the Weibull, estimates are transformed such that
they are in line with the parameterization provided by the \emph{stats} package
like \code{\link{pweibull}}.
}
\details{
When using this method, the approximated confidence intervals for the Weibull
parameters (based on p. 51 of Ralf Mock) can only be estimated for the
following confidence levels:
\itemize{
  \item \code{conf_level} = 0.90,
  \item \code{conf_level} = 0.95,
  \item \code{conf_level} = 0.99.}

If the distribution is not the Weibull, the confidence intervals of the
parameters are calculated using a heteroscedasticity-consistent covariance matrix.
Here it should be said that there is no statistical foundation to calculate the
standard errors for the parameters using \emph{Least Squares} in context of
\emph{Median Rank Regression}. For an accepted statistical method use MLE
(\code{\link{ml_estimation}}).
}
\section{Methods (by class)}{

\describe{
  \item{\code{\link[=rank_regression.cdf_estimation]{cdf_estimation}}}{
    Preferred. Provide the output of \code{\link{estimate_cdf}} directly.
  }
  \item{\code{\link[=rank_regression.default]{default}}}{
    Provide \code{x}, \code{y} and \code{event} manually.
  }
}
}

\references{
\itemize{
    \item Mock, R., Methoden zur Datenhandhabung in
  Zuverlässigkeitsanalysen, vdf Hochschulverlag AG an der ETH Zürich, 1995
    \item Meeker, William Q; Escobar, Luis A., Statistical methods for
  reliability data, New York: Wiley series in probability and statistics, 1998}
}
