% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_estimators.R
\name{rank_regression}
\alias{rank_regression}
\title{Rank Regression for the Weibull distribution}
\usage{
rank_regression(x, y, event, distribution = "weibull", conf_level = 0.95,
  details = TRUE)
}
\arguments{
\item{x}{a numeric vector which consists of lifetime data. Lifetime
data could be every characteristic influencing the reliability of a product,
e.g. operating time (days/months in service), mileage (km, miles), load
cycles.}

\item{y}{a numeric vector which consists of estimated failure probabilities
regarding the lifetime data in \code{x}.}

\item{event}{a vector of binary data (0 or 1) indicating whether unit \emph{i}
is a right censored observation (= 0) or a failure (= 1).}

\item{distribution}{supposed distribution of the random variable. The default
value is \code{"weibull"}.}

\item{conf_level}{confidence level of the interval. The default value is
\code{conf_level = 0.95}}
}
\value{
Returns a list with the following components:
  \itemize{
  \item \code{coefficients} : Where \eqn{\eta} is the scale parameter and
    \eqn{\beta} is the shape parameter
  \item \code{confint} : The estimated confidence intervals of the
    parameters.
  \item \code{r_squared} : Coefficient of determination.}
}
\description{
This method fits an \strong{x on y} regression to the linearized
two-parameter Weibull cdf and is applicable for complete and (multiple) right
censored data. The weibull-specific parameters are estimated in the frequently
used location-scale parametrization and afterwards are transformed such that
they are in line with the parametrization provided by the \emph{stats} package
like \code{\link{pweibull}}.
}
\details{
When using this method, the approximated confidence intervals (based on p. 51
of Ralf Mock) can only be estimated for the following confidence levels:
\itemize{
  \item \code{conf_level} = 0.90,
  \item \code{conf_level} = 0.95,
  \item \code{conf_level} = 0.99.}
}
\examples{
obs   <- seq(10000, 100000, 10000)
state <- c(0, 1, 1, 0, 0, 0, 1, 0, 1, 0)

df_john <- johnson_method(x = obs, event = state)
mrr <- rank_regression(x = df_john$characteristic,
                       y = df_john$prob,
                       event = df_john$status,
                       conf_level = .90)
}
\references{
Mock, R., Methoden zur Datenhandhabung in
  Zuverlässigkeitsanalysen, vdf Hochschulverlag AG an der ETH Zürich, 1995
}
