% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_regression.R
\encoding{UTF-8}
\name{rank_regression}
\alias{rank_regression}
\alias{rank_regression.cdf_estimation}
\title{Rank Regression for Parametric Lifetime Distributions}
\usage{
rank_regression(x, ...)

\method{rank_regression}{cdf_estimation}(
  x,
  distribution = c("weibull", "lognormal", "loglogistic", "normal", "logistic", "sev",
    "weibull3", "lognormal3", "loglogistic3"),
  conf_level = 0.95,
  ...
)
}
\arguments{
\item{x}{Object of class \code{cdf_estimation} returned from
\code{\link{estimate_cdf}}.}

\item{distribution}{Supposed distribution of the random variable.}

\item{conf_level}{Confidence level of the interval. If \code{distribution} is
\code{"weibull"} this must be one of \code{0.9}, \code{0.95} or \code{0.99}.}
}
\value{
Returns a list with the classes \code{"rank_regression"} and
  \code{"model_estimation"} containing the following elements:
  \itemize{
    \item \code{coefficients} : If \code{distribution} is \code{"weibull"}, the
      estimated scale (\eqn{\eta}) and shape (\eqn{\beta}) parameters are provided
      (and additionally the threshold parameter \eqn{\gamma} if \code{distribution}
      is \code{"weibull3"}). For any other distribution, \code{coefficients} is
      identical to \code{loc_sc_params}.
    \item \code{confint} : Only included if \code{distribution} is \code{"weibull"}
      or \code{"weibull3"}. Approximated confidence intervals for \eqn{\eta} and
      \eqn{\beta} (and \eqn{\gamma} if \code{distribution} is \code{"weibull3"}).
    \item \code{loc_sc_params} : Estimated (log-)location-scale parameters.
      Threshold parameter is included if \code{distribution} is \code{"weibull3"},
      \code{"lognormal3"} or \code{"loglogistic3"}.
    \item \code{loc_sc_confint} : Confidence intervals for (log-)location-scale
      parameters. If distribution is \code{"lognormal3"} or \code{"loglogistic3"}
      a confidence interval for the threshold parameter is not computed.
    \item \code{loc_sc_varcov} : Provided, if \code{distribution} is not
      \code{"weibull"} or \code{"weibull3"}. Estimated heteroscedasticity-consistent
      variance-covariance matrix for the (log-)location-scale parameters.
    \item \code{r_squared} : Coefficient of determination.
    \item \code{data} : A tibble with class \code{"cdf_estimation"} returned by
      \code{\link{estimate_cdf}}.
    \item \code{distribution} : Specified distribution.
  }
  If more than one method was specified in \code{\link{estimate_cdf}}, the
  resulting output is a list with class \code{"model_estimation_list"}. In
  this case each list element has classes \code{"rank_regression"} and
  \code{"model_estimation"} and the items listed above, are included.
}
\description{
This function fits an \strong{x on y} regression to a linearized two- or
three-parameter lifetime distribution for complete and (multiple) right
censored data. The parameters are determined in the frequently used
(log-)location-scale parameterization.

For the Weibull, estimates are transformed such that they are in line with the
parameterization provided by the \emph{stats} package (see \link[stats]{Weibull}).
}
\details{
If \code{distribution} is \code{"weibull"} or \code{"weibull3"}, the approximated
confidence intervals for the parameters can only be estimated on the following
confidence levels (see 'References' \emph{(Mock, 1995)}):
\itemize{
  \item \code{conf_level} = 0.90,
  \item \code{conf_level} = 0.95,
  \item \code{conf_level} = 0.99.
}

If the distribution is not the Weibull, the confidence intervals of the
parameters are computed on the basis of a heteroscedasticity-consistent
covariance matrix. Here it should be said that there is no statistical foundation
to determine the standard errors of the parameters using \emph{Least Squares}
in context of \emph{Rank Regression}. For an accepted statistical method use
\link[=ml_estimation]{maximum likelihood}.
}
\examples{
# Reliability data preparation:
## Data for two-parametric model:
data_2p <- reliability_data(
  data = shock,
  x = distance,
  status = status
)

## Data for three-parametric model:
data_3p <- reliability_data(
  data = alloy,
  x = cycles,
  status = status
)

# Probability estimation:
prob_tbl_2p <- estimate_cdf(
  x = data_2p,
  methods = "johnson"
)

prob_tbl_3p <- estimate_cdf(
  x = data_3p,
  methods = "johnson"
)

prob_tbl_mult <- estimate_cdf(
  x = data_3p,
  methods = c("johnson", "kaplan")
)

# Example 1 - Fitting a two-parametric weibull distribution:
rr_2p <- rank_regression(
  x = prob_tbl_2p,
  distribution = "weibull"
)

# Example 2 - Fitting a three-parametric lognormal distribution:
rr_3p <- rank_regression(
  x = prob_tbl_3p,
  distribution = "lognormal3",
  conf_level = 0.99
)

# Example 3 - Fitting a three-parametric loglogistic distribution if multiple
# methods in estimate_cdf were specified:
rr_lists <- rank_regression(
  x = prob_tbl_mult,
  distribution = "loglogistic3",
  conf_level = 0.90
)

}
\references{
\itemize{
    \item Mock, R., Methoden zur Datenhandhabung in
  Zuverlässigkeitsanalysen, vdf Hochschulverlag AG an der ETH Zürich, 1995
    \item Meeker, William Q; Escobar, Luis A., Statistical methods for
  reliability data, New York: Wiley series in probability and statistics, 1998}
}
\seealso{
\code{\link{rank_regression.default}}
}
