% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_regression.R
\encoding{UTF-8}
\name{rank_regression}
\alias{rank_regression}
\alias{rank_regression.cdf_estimation}
\alias{rank_regression.default}
\title{Rank Regression for Parametric Lifetime Distributions}
\usage{
rank_regression(x, ...)

\method{rank_regression}{cdf_estimation}(
  cdf_estimation,
  distribution = c("weibull", "lognormal", "loglogistic", "normal", "logistic", "sev",
    "weibull3", "lognormal3", "loglogistic3"),
  conf_level = 0.95
)

\method{rank_regression}{default}(
  x,
  y,
  event,
  distribution = c("weibull", "lognormal", "loglogistic", "normal", "logistic", "sev",
    "weibull3", "lognormal3", "loglogistic3"),
  conf_level = 0.95
)
}
\arguments{
\item{x}{A numeric vector which consists of lifetime data. Lifetime
data could be every characteristic influencing the reliability of a product,
e.g. operating time (days/months in service), mileage (km, miles), load
cycles.}

\item{distribution}{Supposed distribution of the random variable. The
value can be \code{"weibull"}, \code{"lognormal"}, \code{"loglogistic"},
\code{"normal"}, \code{"logistic"}, \code{"sev"} \emph{(smallest extreme value)},
\code{"weibull3"}, \code{"lognormal3"} or \code{"loglogistic3"}.
Other distributions have not been implemented yet.}

\item{conf_level}{Confidence level of the interval. The default value is
\code{conf_level = 0.95}.}

\item{y}{A numeric vector which consists of estimated failure probabilities
regarding the lifetime data in \code{x}.}

\item{event}{A vector of binary data (0 or 1) indicating whether unit \emph{i}
is a right censored observation (= 0) or a failure (= 1).}

\item{data}{A tibble returned by \code{\link{estimate_cdf}}.}
}
\value{
Returns a list with the following components:
  \itemize{
  \item \code{coefficients} : Provided, if \code{distribution} is \code{"weibull"}
    or \code{"weibull3"}. \eqn{\eta} is the estimated scale and \eqn{\beta}
    the estimated shape parameter. The estimated threshold parameter \eqn{\gamma}
    is available if the three-parametric weibull is used.
  \item \code{confint} : Provided, if \code{distribution} is \code{"weibull"}
    or \code{"weibull3"}. Confidence intervals for \eqn{\eta} and \eqn{\beta}
    (and \eqn{\gamma} if the three-parametric weibull is used).
  \item \code{loc_sc_params} : Estimated location-scale parameters.
    Threshold parameter is provided for \code{"weibull3"}, \code{"lognormal3"}
    and \code{"loglogistic3"}.
  \item \code{loc_sc_confint} : Confidence intervals for location-scale parameters.
    If distribution is \code{"lognormal3"} or \code{"loglogistic3"} a confidence
    interval for the threshold is not computed.
  \item \code{loc_sc_varcov} : Provided, if \code{distribution} is not
    \code{"weibull"} or \code{"weibull3"}. Estimated heteroscedasticity-consistent
    Variance-Covariance matrix of the used location-scale distribution.
  \item \code{r_squared} : Coefficient of determination.}
}
\description{
This method fits an \strong{x on y} regression to a linearized
two- or three-parameter cdf and is applicable for complete and (multiple) right
censored data. The parameters are estimated in the frequently used (log-) location-scale
parameterization. For the Weibull, estimates are transformed such that
they are in line with the parameterization provided by the \emph{stats} package
like \code{\link{pweibull}}.
}
\details{
When using this method, the approximated confidence intervals for the Weibull
parameters (based on p. 51 of Ralf Mock) can only be estimated for the
following confidence levels:
\itemize{
  \item \code{conf_level} = 0.90,
  \item \code{conf_level} = 0.95,
  \item \code{conf_level} = 0.99.}

If the distribution is not the Weibull, the confidence intervals of the
parameters are calculated using a heteroscedasticity-consistent covariance matrix.
Here it should be said that there is no statistical foundation to calculate the
standard errors for the parameters using \emph{Least Squares} in context of
\emph{Median Rank Regression}. For an accepted statistical method use MLE
(\code{\link{ml_estimation}}).
}
\section{Methods (by class)}{
\itemize{
\item \code{cdf_estimation}: Perform the rank regression as part of the
reliability pipeline. \code{\link{estimate_cdf}} returns a data frame of class
\code{"cdf_estimation"}, which contains all information regarding x, y and
event.

\item \code{default}: Provide x, y and event manually.
}}

\examples{
# Example 1: Fitting a two-parameter Weibull:
obs   <- seq(10000, 100000, 10000)
status <- c(0, 1, 1, 0, 0, 0, 1, 0, 1, 0)
reliability_tbl <- reliability_data(x = obs, status = status)

tbl_john <- estimate_cdf(reliability_tbl, "johnson")

mrr <- rank_regression(
  tbl_john,
  distribution = "weibull",
  conf_level = .90
)

# Example 2: Fitting a three-parameter Weibull:
# Alloy T7987 dataset taken from Meeker and Escobar(1998, p. 131)
cycles   <- c(300, 300, 300, 300, 300, 291, 274, 271, 269, 257, 256, 227, 226,
              224, 213, 211, 205, 203, 197, 196, 190, 189, 188, 187, 184, 180,
              180, 177, 176, 173, 172, 171, 170, 170, 169, 168, 168, 162, 159,
              159, 159, 159, 152, 152, 149, 149, 144, 143, 141, 141, 140, 139,
              139, 136, 135, 133, 131, 129, 123, 121, 121, 118, 117, 117, 114,
              112, 108, 104, 99, 99, 96, 94)
status <- c(rep(0, 5), rep(1, 67))

reliability_tbl <- reliability_data(x = cycles, status = status)

tbl_john <- estimate_cdf(reliability_tbl, "johnson")
mrr <- rank_regression(
  tbl_john,
  distribution = "weibull3",
  conf_level = .90
)

}
\references{
\itemize{
    \item Mock, R., Methoden zur Datenhandhabung in
  Zuverlässigkeitsanalysen, vdf Hochschulverlag AG an der ETH Zürich, 1995
    \item Meeker, William Q; Escobar, Luis A., Statistical methods for
  reliability data, New York: Wiley series in probability and statistics, 1998}
}
