% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{reliability_data}
\alias{reliability_data}
\title{Reliability data}
\usage{
reliability_data(data, x, event, id)
}
\arguments{
\item{data}{A data frame.}

\item{x}{The name of a column holding lifetime data. Lifetime data could
be every characteristic influencing the reliability of a product, e.g.
operating time (days/months in service), mileage (km, miles), load cycles.}

\item{event}{The name of a column holding binary data (0 or 1) indicating
whether unit \emph{i} is a right censored observation (= 0) or a failure (= 1).}

\item{id}{The name of a column holding the identification for every unit.}
}
\description{
This function prepares a dataset for the reliability pipeline.
}
\examples{
# Alloy T7987 dataset taken from Meeker and Escobar(1998, p. 131)
cycles   <- c(300, 300, 300, 300, 300, 291, 274, 271, 269, 257, 256, 227, 226,
              224, 213, 211, 205, 203, 197, 196, 190, 189, 188, 187, 184, 180,
              180, 177, 176, 173, 172, 171, 170, 170, 169, 168, 168, 162, 159,
              159, 159, 159, 152, 152, 149, 149, 144, 143, 141, 141, 140, 139,
              139, 136, 135, 133, 131, 129, 123, 121, 121, 118, 117, 117, 114,
              112, 108, 104, 99, 99, 96, 94)
state <- c(rep(0, 5), rep(1, 67))
id <- "XXXXXX"

df <- data.frame(cycles = cycles, state = state, id = id)

data <- reliability_data(df, x = cycles, event = state, id = id)

}
