% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{reliability_data}
\alias{reliability_data}
\title{Reliability data}
\usage{
reliability_data(data, x, event, id)
}
\arguments{
\item{data}{A data frame.}

\item{x, event, id}{Column names of columns containing reliability information.}
}
\description{
This function prepares a dataset for the reliability pipeline.
}
